% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absmax.r
\name{absmax}
\alias{absmax}
\title{Maximum absolute value of a vector}
\usage{
absmax(x)
}
\arguments{
\item{x}{vector of numbers}
}
\value{
Maximum of the absolute value of a vector
}
\description{
Generates the maximum of the absolute value of a numeric vector.
}
\details{
Version 1.1, 3/29/2021
}
\examples{
r50 <- rnorm(50)
r50.absmax <- absmax(r50)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{monsell.brian@gmail.com}
}
