% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_outlier_lines.r
\name{add_outlier_lines}
\alias{add_outlier_lines}
\title{add lines for outliers}
\usage{
add_outlier_lines(
  this_p = NULL,
  seas_obj = NULL,
  line_color = c("red", "blue", "orangered", "green", "steelblue", "blue"),
  this_palette = "Dark2",
  this_line_type = c("dashed", "dotdash", "dashed", "twodash", "dotdash", "dotdash")
)
}
\arguments{
\item{this_p}{A \code{ggplot} object of a time series plot.
This is a required entry.}

\item{seas_obj}{\code{seas} object generated from a call of \code{seas} on a single time series.
This is a required entry.}

\item{line_color}{Character array of length 6; color used for different outliers, with the order being
\code{'ao', 'ls', 'tc', 'so', 'rp', 'tls'}.
Default is the \code{RColorBrewer} palette \code{"Dark2"}.}

\item{this_palette}{Character string; default \code{RColorBrewer} palette.
Deault is \code{"Dark2"}.}

\item{this_line_type}{Character array of length 6; Line type used for different outliers, with the order being
\code{'ao', 'ls', 'tc', 'so', 'rp', 'tls'}.
Default is \code{c('dashed', 'dotdash', 'dashed', 'twodash', 'dotdash', 'dotdash')}.}
}
\value{
Revised \code{ggplot} object with lines for outliers added.
}
\description{
add lines for outliers to a ggplot plot object
}
\details{
Version 3.2, 9/5/2024
}
\examples{
air_seas <- 
   seasonal::seas(AirPassengers, arima.model = "(0 1 1)(0 1 1)", x11 = "")
air_df   <- 
   data.frame(date = tsbox::ts_df(AirPassengers)$time, 
              ori = as.double(seasonal::original(air_seas)),  
              sa = as.double(seasonal::final(air_seas)), 
              trend = as.double(seasonal::trend(air_seas)))
this_x11_p <- ggplot2::ggplot() + 
  ggplot2::geom_line(mapping = ggplot2::aes(x=date, y = ori), 
                     color = "grey", data = air_df) +
  ggplot2::geom_line(mapping = ggplot2::aes(x=date, y = sa), 
                     color="steelblue", linetype="twodash", data = air_df) + 
  ggplot2::geom_line(mapping = ggplot2::aes(x=date, y = trend), 
                     color="darkred", linetype="twodash", data = air_df) + 
  ggplot2::labs(
    title = "Airline Passenger X-11 Seasonal Adjustment",
    subtitle = NULL,
    x = "Time",
    y = "Airline Passengers")
this_x11_p_with_outlier_lines <- 
    add_outlier_lines(this_x11_p, air_seas, this_palette = "Paired",
                      this_line_type = rep("dotted", 6))
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{monsell.brian@gmail.com}
}
