% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_recession_shade.r
\name{add_recession_shade}
\alias{add_recession_shade}
\title{Add shading for NBER recession dates}
\usage{
add_recession_shade(
  this_p = NULL,
  shade_color = "lightblue1",
  shade_alpha = 0.2
)
}
\arguments{
\item{this_p}{\code{ggplot} object of a time series plot.
This is a required entry.}

\item{shade_color}{Character scalar; shading for recession region.
Default is \code{"pink"}.}

\item{shade_alpha}{numeric scalar; controls the intensity of the shading. Default is \code{0.2}.}
}
\value{
\code{ggplot} object with shading for recession added.
}
\description{
Add shading for US NBER recession dates ro ggplot plot object.
}
\details{
Version 3.2, 5/6/2024
}
\examples{
air_seas <- 
   seasonal::seas(AirPassengers, arima.model = "(0 1 1)(0 1 1)", x11 = "")
air_df   <- 
   data.frame(date = tsbox::ts_df(AirPassengers)$time, 
              ori = as.double(seasonal::original(air_seas)),  
              sa = as.double(seasonal::final(air_seas)), 
              trend = as.double(seasonal::trend(air_seas)))
this_p <- ggplot2::ggplot(air_df) +
  ggplot2::geom_line(ggplot2::aes(x=date, y = ori), color = "grey") + 
  ggplot2::geom_line(ggplot2::aes(x=date, y = sa), color="steelblue", linetype="twodash") + 
  ggplot2::geom_line(ggplot2::aes(x=date, y = trend), color="darkred", linetype="twodash") + 
  ggplot2::labs(
    title = "Airline Passenger X-11 Seasonal Adjustment",
    subtitle = NULL,
    x = "Time",
    y = "Airline Passengers")
this_p_with_recession_shading <-
    add_recession_shade(this_p, shade_color = "steelblue")
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{monsell.brian@gmail.com}
}
