% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_blind_palette.r
\name{color_blind_palette}
\alias{color_blind_palette}
\title{Color-blind friendly color palette}
\usage{
color_blind_palette(with_grey = TRUE, brewer_palette = NULL)
}
\arguments{
\item{with_grey}{Logical scalar; whether color blind pallate contains \code{'grey'},
otherwise the palette contains \code{black}. Default is TRUE.}

\item{brewer_palette}{Character string; a \code{RColorBrewer} palette.
There is no default - must be a color-blind friendly palette.}
}
\value{
Vector of hexadecimal color codes that form a color palette that can be distinguished by color-blind people.
}
\description{
Color palettes that can be used that can be distinguished by color-blind people (either from \code{RColorBrewer}
or Cookbook for R - Colors (ggplot2)).
}
\details{
Version 2.2, 8/2/2024
}
\examples{
this_color_blind <- color_blind_palette(FALSE, brewer_palette = "Dark2")
}
\references{
\url{https://CRAN.R-project.org/package=RColorBrewer}, \url{http://www.cookbook-r.com/Graphs/Colors_(ggplot2)/http://www.cookbook-r.com/Graphs/Colors_(ggplot2)/}
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{monsell.brian@gmail.com}
}
