% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_identify_acf.r
\name{convert_identify_acf}
\alias{convert_identify_acf}
\title{Convert matrix of ACFs or PACFs generated by X-13ARIMA-SEATS \code{identify} spec to a list object}
\usage{
convert_identify_acf(seas_obj = NULL, this_plot = "iac")
}
\arguments{
\item{seas_obj}{\code{seas} object generated from a call of \code{seas} on a single time series
This is a required entry.}

\item{this_plot}{Character string; three character code for the type of plot to be generated.
Allowed entries are \code{"iac"} (sample autocorrelation function, default),
\code{"ipf"} (sample partial autocorrelation function).}
}
\value{
A list of matrices of ACF or PACFs produced for different orders of differncing.
The list entries are named based on the orders of differencing (\code{d0sd0} denotes
no regular difference, no seasonal difference, \code{d1sd0} denotes
one regular difference, no seasonal difference, etc.)
}
\description{
Generates a list of the ACF or PACF generated by the \code{identify} spec
}
\details{
Version 2.0, 10/23/2024
}
\examples{
shoes_identify_seas <- 
   seasonal::seas(shoes2008, 
                  identify.diff = c(0, 1),
                  identify.sdiff = c(0, 1),
                  identify.save = c("iac", "ipc"),
                  arima.model = "(0 1 1)(0 1 1)", 
                  transform.function = "log", 
                  forecast.maxlead = 36,
			     check.maxlag = 36,
				     check.acflimit = 1.96,
				     check.qlimit = 0.01,
                  check.print = c( 'pacf', 'pacfplot' ))
shoes_identify_acf_list <- 
    convert_identify_acf(shoes_identify_seas, 'iac')
shoes_identify_pacf_list <- 
    convert_identify_acf(shoes_identify_seas, 'ipc')
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
