% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_spectrum_code.r
\name{convert_spectrum_code}
\alias{convert_spectrum_code}
\title{Convert spectrum code to visual peak code}
\usage{
convert_spectrum_code(this_spectrum = "sp0")
}
\arguments{
\item{this_spectrum}{Character string; three character code for the X-13 spectrum to be generated.
Allowed entries are \code{"sp0"} (modified original series),
\code{"sp1"} (modified X-11 seasonally adjusted series),
\code{"sp2"} (modified X-11 irregular),
\code{"s1s"} (modified SEATS seasonally adjusted series),
\code{"s2s"} (modified SEATS irregular),
\code{"is0"} (modified composite series),
\code{"is1"} (modified indirect seasonally adjusted series),
\code{"is2"} (modified indirect irregular), \code{spr} (model residuals), or
\code{"ser"} (extended residuals).  Default: \code{"sp0"}.}
}
\value{
Text for spectrum associated with code used in \code{plot_spectrum} function.
If improper \code{this_spectrum} specified, function will return \code{NULL}.
}
\description{
Convert the three character code used by the \code{plot_spectrum} function so that it can be used with the
\code{visual_sig_peaks} function from the \code{blsplot} package.
}
\details{
Version 1.4, 5/6/2024
}
\examples{
this_code <- convert_spectrum_code("sp2")
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{monsell.brian@gmail.com}
}
