% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employment_data_mts.r
\docType{data}
\name{employment_data_mts}
\alias{employment_data_mts}
\title{US Unemployment Series, four main components in an mts object}
\format{
An mts object with 4 time series elements in four columns:
\describe{
\item{n2000013}{Employed Males 16-19}
\item{n2000014}{Employed Females 16-19}
\item{n2000025}{Employed Males 20+}
\item{n2000026}{Employed Females 20+}
}
}
\usage{
employment_data_mts
}
\description{
An mts object of the four main components of US Employment expressed as time series objects that end in December, 2022
}
\keyword{datasets}
