% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employment_list.r
\docType{data}
\name{employment_list}
\alias{employment_list}
\title{US Employment Series, four main components in a list object}
\format{
A list object with 4 time series elements:
\describe{
\item{n2000013}{Employed Males 16-19}
\item{n2000014}{Employed Females 16-19}
\item{n2000025}{Employed Males 20+}
\item{n2000026}{Employed Females 20+}
}
}
\usage{
employment_list
}
\description{
A list object of the four main components of US Employment expressed as time series objects that end in December, 2022
}
\keyword{datasets}
