% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_range_from_ggplot.r
\name{extract_range_from_ggplot}
\alias{extract_range_from_ggplot}
\title{Extract range of data from ggplot object}
\usage{
extract_range_from_ggplot(this_p = NULL)
}
\arguments{
\item{this_p}{\code{ggplot} object of a time series plot.
This is a required entry.}
}
\value{
Vector of length 2 with the range of the data used to generate a given
\code{ggplot} object of a time series plot.
}
\description{
Computes the range of all data plotted in given ggplot object
}
\details{
Version 1.2, 10/19/2023
}
\examples{
air_seas <- 
   seasonal::seas(AirPassengers, arima.model = "(0 1 1)(0 1 1)", x11 = "")
air_df   <- 
   data.frame(date = tsbox::ts_df(AirPassengers)$time, 
              ori = as.double(seasonal::original(air_seas)),  
              sa = as.double(seasonal::final(air_seas)), 
              trend = as.double(seasonal::trend(air_seas)))
air_p <- ggplot2::ggplot(air_df, ggplot2::aes(x=date)) +
  ggplot2::geom_line(ggplot2::aes(y = ori), color = "grey") + 
  ggplot2::geom_line(ggplot2::aes(y = trend), color="darkred", 
                     linetype="twodash") + 
  ggplot2::labs(
    title = "Airline Passenger X-11 Trend Component",
    subtitle = NULL,
    x = "Time",
    y = "Airline Passengers")
air_short_p <- plot_date_span(air_p, "1-1-1956", "1-1-1961")
air_short_range <- extract_range_from_ggplot(air_short_p)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{monsell.brian@gmail.com}
}
