% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_alt_text.r
\name{generate_alt_text}
\alias{generate_alt_text}
\title{Generate alt text for ggplot graphs}
\usage{
generate_alt_text(
  gg_object = NULL,
  chart_type = NULL,
  data_type = NULL,
  reason_text = NULL,
  short_alt = FALSE
)
}
\arguments{
\item{gg_object}{\code{ggplot} object from which alt text will be generated.
Required entry if \code{short_alt = FALSE}.}

\item{chart_type}{character scalar telling what type of plot is used in \code{gg_object}.
This is a required entry.}

\item{data_type}{character scalar detailing what data is used in \code{gg_object}.
This is a required entry.}

\item{reason_text}{character scalar detailing the reason \code{gg_object} is plotted.
This is a required entry.}

\item{short_alt}{logical scalar if TRUE BrailleR text will not be appended to the alt text.
Default is \code{FALSE}.}
}
\value{
generate alt text for plot produced by \code{gg_object}
}
\description{
Generates alt text for ggplot graph objects using the \code{BrailleR} package and adding
text suggested by Amy Casale in her article "Writing Alt Text for Data Visualization".
}
\details{
Version 2.1, 5/6/2024
}
\examples{
air_seas <- seasonal::seas(AirPassengers, arima.model = "(0 1 1)(0 1 1)", x11="")
air_df   <- 
   data.frame(date = tsbox::ts_df(AirPassengers)$time, 
              ori = as.double(seasonal::original(air_seas)), 
              sa = as.double(seasonal::final(air_seas)), 
              trend = as.double(seasonal::trend(air_seas)))
air_p <- ggplot2::ggplot(air_df, ggplot2::aes(x=date)) +
  ggplot2::geom_line(ggplot2::aes(y = ori), color = "grey") + 
  ggplot2::geom_line(ggplot2::aes(y = sa), 
                     color="steelblue", linetype="twodash") + 
  ggplot2::geom_line(ggplot2::aes(y = trend), 
                     color="darkred", linetype="dotdash") + 
  ggplot2::ggtitle("Airline Passenger X-11 Seasonal Adjustment")
air_alt_text <- 
   generate_alt_text(air_p, 
                     "Time series plot", 
                     "International Airline Passengers time series",
                     "compare seasonal adjustment and trend to original series")
}
\references{
\url{https://CRAN.R-project.org/package=BrailleR} and   \url{https://medium.com/nightingale/writing-alt-text-for-data-visualization-2a218ef43f81}
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{monsell.brian@gmail.com}
}
