% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_auto_outlier_string.r
\name{get_auto_outlier_string}
\alias{get_auto_outlier_string}
\title{Get automatic outlier names}
\usage{
get_auto_outlier_string(seas_obj = NULL)
}
\arguments{
\item{seas_obj}{A \code{seas} object for a single series generated from the
\code{seasonal} package. This is a required entry.}
}
\value{
Character string containing a summary of the outliers identified in the
regARIMA model. If no regressors or automatic outliers in the model, the routine
will return a blank character.
}
\description{
Get the names of outliers identified in the \code{seas} object for a single series.
}
\details{
Version 3.0, 5/14/2024
}
\examples{
air_seas <- seasonal::seas(AirPassengers, arima.model = "(0 1 1)(0 1 1)", x11="")
this_auto_outlier <- get_auto_outlier_string(air_seas)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
