% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_month_index.R
\name{get_month_index}
\alias{get_month_index}
\title{Generate index of month abbreviation}
\usage{
get_month_index(this_month_string)
}
\arguments{
\item{this_month_string}{Character string; 3 character abbreviation of month}
}
\value{
Index of month - 1 for 'Jan', 2 for 'Feb', etc.
}
\description{
Process string of month abbrev to return a numeric index.
}
\details{
Version 2.3, 9/18/2020
}
\examples{
thisOtl <- 'AO2015.Jan'
thisCode <- 'AO'
thisPerChar <- substr(thisOtl,nchar(thisCode)+6,nchar(thisOtl))
thisPerIndex <- get_month_index(thisPerChar)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{monsell.brian@gmail.com}
}
