% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reg_string.r
\name{get_reg_string}
\alias{get_reg_string}
\title{Get names of regressors}
\usage{
get_reg_string(seas_obj = NULL, xreg_names = NULL)
}
\arguments{
\item{seas_obj}{\code{seas} object generated by the \code{seasonal} package for a single series.
This is a required entry.}

\item{xreg_names}{Character vector with names of user defined regressors used in model.
Default is \code{NULL}, no user defined regressors. Number of names in this vector should
match number of user-defined regressors; if not, a warning message will be produced.}
}
\value{
Character string containing a summary of the regressors in the regARIMA model.
If no regressors in the model, the routine will return a blank character.
}
\description{
Generate string of names for the regressors used in the model fit for a given series.
}
\details{
Version 2.9, 5/14/2024
}
\examples{
air_seas <- seasonal::seas(AirPassengers, slidingspans = "", 
                  transform.function = "log", 
                  x11 = "", forecast.maxlead=36, 
                  check.print = c( "pacf", "pacfplot" ))
air_reg <- get_reg_string(air_seas)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{monsell.brian@gmail.com}
}
