% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_acf.r
\name{plot_acf}
\alias{plot_acf}
\title{Generate ACF plot of the regARIMA model residuals.}
\usage{
plot_acf(
  seas_obj = NULL,
  this_plot = "acf",
  acf_range = NULL,
  add_ci = TRUE,
  main_title = "ACF Plot",
  sub_title = NULL,
  this_x_label = "Lag",
  this_y_label = "ACF",
  do_grid = FALSE,
  do_background = FALSE,
  acf_color = "steelblue"
)
}
\arguments{
\item{seas_obj}{\code{seas} object generated from a call of \code{seas} on a single time series
This is a required entry.}

\item{this_plot}{Character string; three character code for the type of plot to be generated.
Allowed entries are \code{"acf"} (sample autocorrelation function, default),
\code{"pcf"} (sample partial autocorrelation function),
\code{"ac2"} (squared autocorrelation function).}

\item{acf_range}{Range of values you wish the acf plot to be plotted over.
Default is range of the series.}

\item{add_ci}{Logical scalar; indicates if confidence intervals are added to the plot.
Default is confidence intervals are added.}

\item{main_title}{Title for the plot.
Default is character string \code{'ACF Plot'}.}

\item{sub_title}{Subtitle for the plot. Default is \code{NULL}, or no subtitle.}

\item{this_x_label}{Label for X axis. Default is \code{"Lags"}.}

\item{this_y_label}{Label for Y axis. Default is \code{"ACF"}.}

\item{do_grid}{Logical scalar; indicates if certain plots will have grid lines.
Default is no grid lines.}

\item{do_background}{Logical scalar; indicates grey background included in plot.
Default is no grey background;}

\item{acf_color}{Color used for lines in ACF plot.
Default is \code{"steelblue"}.}
}
\value{
A \code{ggplot} object that produces an ACF, PACF, or squared ACF plot
of the regARIMA residuals.
}
\description{
Generate ACF, PACF, or squared ACF plot of the regARIMA model residuals.
}
\details{
Version 2.2, 10/23/2024
}
\examples{
ukgas_x11_seas <- 
   seasonal::seas(UKgas, series.period = 4, 
                  arima.model = "(0 1 1)(0 1 1)", 
                  transform.function = "log", 
                  forecast.maxlead = 20,
                  x11.seasonalma = "s3x5", 
                  check.print = c( 'pacf', 'pacfplot' ),
                  check.maxlag = 12,
                  check.save = c("acf", "pcf", "ac2"))
p_ukgas_acf   <- 
   plot_acf(ukgas_x11_seas, this_plot = "acf",
            main_title = "UK Gas Model Residual ACF",
            acf_color = "darkblue")
p_ukgas_pacf   <- 
   plot_acf(ukgas_x11_seas, this_plot = "pcf",
            main_title = "UK Gas Model Residual PACF",
            acf_color = "darkblue")
p_ukgas_acf2   <- 
   plot_acf(ukgas_x11_seas, this_plot = "ac2",
            main_title = "UK Gas Model Squared ACF",
            acf_color = "darkblue")
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
