% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_acf_and_pacf.r
\name{plot_acf_and_pacf}
\alias{plot_acf_and_pacf}
\title{Generate ACF and PACF plot of the regARIMA model residuals.}
\usage{
plot_acf_and_pacf(
  seas_obj = NULL,
  add_ci = TRUE,
  overall_title = NULL,
  acf_title = "ACF Plot",
  pacf_title = "PACF Plot",
  this_x_label = "Lag",
  this_y_label = c("ACF", "PACF"),
  this_x_axis_breaks = NULL,
  do_grid = FALSE,
  do_background = FALSE,
  acf_color = "steelblue"
)
}
\arguments{
\item{seas_obj}{\code{seas} object generated from a call of \code{seas} on a single time series
This is a required entry.}

\item{add_ci}{Character scalar; Overall title for the plot.
Default is confidence intervals are added.}

\item{overall_title}{Overall title for the combined plot.
Default is \code{NULL}, which doesn't produce an overall title.}

\item{acf_title}{Subtitle for the ACF Plot.
Default is character string \code{'ACF Plot'}.}

\item{pacf_title}{Subtitle for the PACF Plot.
Default is character string \code{'PACF Plot'}.}

\item{this_x_label}{Label for X axis. Default is \code{"Lags"}.}

\item{this_y_label}{Character vector of length two. Labels for each Y axis.
Default is \code{c("ACF", "PACF")}.}

\item{this_x_axis_breaks}{Numeric vector; sets the values for the x-axis.
Default uses the value of \code{this_frequency} to set x-axis.}

\item{do_grid}{Logical scalar; indicates if certain plots will have grid lines.
Default is no grid lines.}

\item{do_background}{Logical scalar; indicates grey background included in plot.
Default is no grey background;}

\item{acf_color}{Color used for lines in ACF plot.
Default is \code{"steelblue"}.}
}
\value{
A \code{ggplot} object that produces an ACF and PACF plot for the regARIMA
model residuals.
}
\description{
Generate a single plot with ACF and PACF of the regARIMA model residuals.
}
\details{
Version 3.1, 11/6/2024
}
\examples{
ukgas_x11_seas <- 
   seasonal::seas(UKgas, series.period = 4, 
                  arima.model = "(0 1 1)(0 1 1)", 
                  transform.function = "log", 
                  forecast.maxlead = 20,
                  x11.seasonalma = "s3x5", 
                  check.print = c( 'pacf', 'pacfplot' ),
                  check.maxlag = 12,
                  check.save = c("acf", "pcf", "ac2"))
p_ukgas_acf_and_pacf   <- 
   plot_acf_and_pacf(ukgas_x11_seas, 
            overall_title = "UK Gas Model",
            acf_color = "darkblue")
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
