% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_acf_and_pacf_identify.r
\name{plot_acf_and_pacf_identify}
\alias{plot_acf_and_pacf_identify}
\title{Generate ACF and PACF plot of the regARIMA model residuals.}
\usage{
plot_acf_and_pacf_identify(
  seas_obj = NULL,
  this_diff = 0,
  this_sdiff = 0,
  add_ci = TRUE,
  overall_title = NULL,
  acf_title = "ACF Plot",
  pacf_title = "PACF Plot",
  this_x_label = "Lag",
  this_y_label = c("ACF", "PACF"),
  this_x_axis_breaks = NULL,
  do_grid = FALSE,
  do_background = FALSE,
  acf_color = "steelblue"
)
}
\arguments{
\item{seas_obj}{\code{seas} object generated from a call of \code{seas} on a single time series
This is a required entry.}

\item{this_diff}{Integer scalar; one of the regular differences specified in the
\code{diff} argument of the \code{identify} spec. Default is \code{0}.}

\item{this_sdiff}{Integer vector; one of the seasonal differences specified in the
\code{sdiff} argument of the \code{identify} spec. Default is \code{0}.}

\item{add_ci}{Character scalar; Overall title for the plot.
Default is confidence intervals are added.}

\item{overall_title}{Overall title for the combined plot.
Default is a text string showing the orders of differencing selected.}

\item{acf_title}{Subtitle for the ACF Plot.
Default is character string \code{'ACF Plot'}.}

\item{pacf_title}{Subtitle for the PACF Plot.
Default is character string \code{'PACF Plot'}.}

\item{this_x_label}{Label for X axis. Default is \code{"Lags"}.}

\item{this_y_label}{Character vector of length two. Labels for each Y axis.
Default is \code{c("ACF", "PACF")}.}

\item{this_x_axis_breaks}{Numeric vector; sets the values for the x-axis.
Default uses the value of \code{this_frequency} to set x-axis.}

\item{do_grid}{Logical scalar; indicates if certain plots will have grid lines.
Default is no grid lines.}

\item{do_background}{Logical scalar; indicates grey background included in plot.
Default is no grey background;}

\item{acf_color}{Color used for lines in ACF plot.
Default is \code{"steelblue"}.}
}
\value{
A \code{ggplot} object that produces an ACF and PACF plot for the original
series with user-specified orders of differencing.
}
\description{
Generate a single plot with ACF and PACF of the regARIMA model residuals.
}
\details{
Version 3.0, 11/5/2024
}
\examples{
shoes_identify_seas <- 
   seasonal::seas(shoes2008, 
                  identify.diff = c(0, 1),
                  identify.sdiff = c(0, 1),
                  identify.save = c("iac", "ipc"),
                  arima.model = "(0 1 1)(0 1 1)", 
                  transform.function = "log", 
                  forecast.maxlead = 36,
			     check.maxlag = 36,
				     check.acflimit = 1.96,
				     check.qlimit = 0.01,
                  check.print = c( 'pacf', 'pacfplot' ))
p_shoes_acf_and_pacf_identify_d0_sd0   <- 
   plot_acf_and_pacf_identify(shoes_identify_seas, 
            overall_title = "US Shoe Sales - No Differencing",
            acf_color = "darkblue")
p_shoes_acf_and_pacf_identify_d1_sd1   <- 
   plot_acf_and_pacf_identify(shoes_identify_seas, 
           this_diff = 1,
		   this_sdiff = 1,
            overall_title = "US Shoe Sales - Regular and Seasonal Differencing",
            acf_color = "darkblue")
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
