% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all_trend_lags.r
\name{plot_all_trend_lags}
\alias{plot_all_trend_lags}
\title{History Plot of All Trend Revisions}
\usage{
plot_all_trend_lags(
  seas_obj = NULL,
  this_trend_name = NULL,
  main_title = "All Trend Revision Plot",
  sub_title = NULL,
  this_y_label = NULL,
  this_x_label = "Time",
  do_grid = FALSE,
  do_background = FALSE,
  base_color = "darkblue",
  whisker_color = "darkgrey"
)
}
\arguments{
\item{seas_obj}{\code{seas} object generated from a call of \code{seas} on a single time series
This is a required entry.}

\item{this_trend_name}{Character string; name of time series. No default.}

\item{main_title}{Title for the plot. Default is character string \code{'Trend History Plot'}.}

\item{sub_title}{Subtitle for the plot. Default is \code{NULL}.}

\item{this_y_label}{Label for Y axis. Default is \code{"Ratio"}.}

\item{this_x_label}{Label for X axis. Default is \code{"Time"}.}

\item{do_grid}{Logical scalar; indicates if certain plots will have grid lines.
Default is no grid lines.}

\item{do_background}{Logical scalar; indicates grey background included in plot.
Default is no grey background;}

\item{base_color}{Character scalar for plot of the initial trend.
Default is \code{"darkblue"}.}

\item{whisker_color}{Character scalar for color used for lines representing
lagged trend estimates in the trend history plot.
Default is \code{"darkgrey"}}
}
\value{
A \code{ggplot} object that produces a history plot of the trend component.
The \code{seas} object in the \code{seas_obj} argument must contain output for a
revisions history analysis for trends with \code{trendlags}
set to some values - these lags should be in sequence, such as
\code{history.trendlags = 1:4}.
}
\description{
Generates a ggplot2 object of estimates from a revisions history of a trend
component for a given series, incorporating all trend lag revisions.
}
\details{
Version 1.8, 9/5/20244
}
\examples{
shoes_seas <-
    seasonal::seas(shoes2008, x11.save = "d13",
       x11 = "", transform.function = "log",
       check.print = c("none", "+acf", "+acfplot", "+normalitytest"),
       regression.aictest = c("td", "easter"),
	  regression.save = c("td", "hol"),
       outlier.types = "all",
       arima.model = "(0 1 1)(0 1 1)",
       forecast.maxlead = 60,
	   	  history.estimates = c("sadj", "sadjchng", "trend"),
	      history.trendlags = 1:4,
	      history.print = "all",
	      history.save = c("tre", "trr"))
p_shoes_sa_history <- 
    plot_all_trend_lags(shoes_seas, this_trend_name = "shoes", 
        main_title = "All Trend Revisions, Lags 1 to 4",
        sub_title = "US Retail Sales of Shoes")
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{monsell.brian@gmail.com}
}
