% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cpgram_resid.R
\name{plot_cpgram_resid}
\alias{plot_cpgram_resid}
\title{Generate cumulative periodogram of the regARIMA residuals}
\usage{
plot_cpgram_resid(
  seas_obj = NULL,
  main_title = "Cumulative periodogram",
  this_palette = NULL
)
}
\arguments{
\item{seas_obj}{\code{seas} object generated from a call of \code{seas} on a single time series.
This is a required entry.}

\item{main_title}{Title for the plot. Default is character string \code{'Cumulative periodogram'}.}

\item{this_palette}{Color used for lines in plot.  Default is a color-blind friendly palette
generated by the function \code{color_blind_palette(FALSE)}}
}
\value{
Generates a \code{ggplot} object of the cumulative periodogram of the regARIMA residuals.
Diagnostic information is included in the plot subheader.
}
\description{
Generates a plot of the cumulative periodogram of the regARIMA residuals.
}
\details{
Version 1.6, 7/1/2024
}
\examples{
air_seas <- 
   seasonal::seas(AirPassengers, transform.function= "log", 
                  arima.model = "(0 1 1)(0 1 1)", check.save = "acf")
plot_cpgram_resid_air <-
     plot_cpgram_resid(air_seas, 
                       main_title = "Cumulative periodogram for Airline Passenger Residuals")
}
