% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_date_span.r
\name{plot_date_span}
\alias{plot_date_span}
\title{Plot a span of data}
\usage{
plot_date_span(
  this_p = NULL,
  this_start_span = NULL,
  this_end_span = NULL,
  this_date_breaks = "1 year",
  this_date_format = "\%Y",
  reset_y_limit = FALSE
)
}
\arguments{
\item{this_p}{A \code{ggplot} object of a time series plot.
This is a required entry.}

\item{this_start_span}{Character scalar with the date of the start of the span to be plotted.
This is a required entry.}

\item{this_end_span}{Character scalar with the date of the end of the span to be plotted.
This is a required entry.}

\item{this_date_breaks}{Character scalar with the interval for tic marks on the x-axis.
Default is \code{"1 year"}.}

\item{this_date_format}{Character scalar with the format used for the date label on the x-axis.}

\item{reset_y_limit}{Logical scalar that if TRUE, the range of the y-axis is reset. Default is \code{FALSE}.}
}
\value{
A \code{ggplot} object that produces a subplot of the submitted plot.
}
\description{
Shortens the time span of an existing time series \code{ggplot} object by limiting the
X axis to user specified dates.
}
\details{
Version 2.2, 8/29/2024
}
\examples{
air_seas <- 
   seasonal::seas(AirPassengers, arima.model = "(0 1 1)(0 1 1)", x11 = "")
air_df   <- 
   data.frame(date = tsbox::ts_df(AirPassengers)$time, 
              ori = as.double(seasonal::original(air_seas)),  
              sa = as.double(seasonal::final(air_seas)), 
              trend = as.double(seasonal::trend(air_seas)))
air_p <- ggplot2::ggplot(air_df, ggplot2::aes(x=date)) +
  ggplot2::geom_line(ggplot2::aes(y = ori), color = "grey") + 
  ggplot2::geom_line(ggplot2::aes(y = trend), 
                     color="darkred", 
                     linetype="twodash") + 
  ggplot2::labs(
    title = "Airline Passenger X-11 Trend Component",
    subtitle = NULL,
    x = "Time",
    y = "Airline Passengers")
air_short_p <- 
   plot_date_span(air_p, "1-1-1956", "1-1-1962", reset_y_limit = TRUE)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
