% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_double_spectrum.R
\name{plot_double_spectrum}
\alias{plot_double_spectrum}
\title{Generate double spectrum plot of the original and seasonally adjusted series.}
\usage{
plot_double_spectrum(
  seas_obj = NULL,
  xaxis_bls = TRUE,
  main_title = "AR Spectrum",
  series_name = NULL,
  this_color = c("darkblue", "darkgreen"),
  this_median_color = c("blue", "green"),
  this_freq_color = c("steelblue", "forestgreen"),
  this_peak_color = c("violet", "brown")
)
}
\arguments{
\item{seas_obj}{\code{seas} object generated from a call of \code{seas} on a single time series
This is a required entry.}

\item{xaxis_bls}{Logical scalar; indicates if x-axis of spectral plot will be frequency by month
rather than the actual frequencies. Default sets x-axis to frequency by month.}

\item{main_title}{Character string; main title of plot.  Default is  \code{'AR Spectrum'}.}

\item{series_name}{Character scalar; name of the time series used in \code{seas_obj}.
Used as the label of the Y-axis if specified.}

\item{this_color}{Character vector of length 2. Colors used for original and seasonally adjusted spectrum in plot.
Defaults are \code{c("darkblue", "darkgreen")}.}

\item{this_median_color}{Character vector of length 2. Colors used for medians of the original and
seasonally adjusted spectrum, respectively. Defaults are \code{c("blue", "green")}.}

\item{this_freq_color}{Character vector of length 2. Colors used for seasonal and trading day frequencies,
respectively. Defaults are \code{c("steelblue", "forestgreen")}.}

\item{this_peak_color}{Character vector of length 2. Colors used for peaks at seasonal and trading day
frequencies, respectively. Defaults are \code{c("violet", "brown")}.}
}
\value{
\code{ggplot} object of spectrum of original series and seasonally adjusted series on same axis.
}
\description{
Generate plot of spectrum of original series and seasonally adjusted series on same axis.
}
\details{
Version 2.5, 7/1/2024
}
\examples{
air_seas <- seasonal::seas(AirPassengers, arima.model = "(0 1 1)(0 1 1)", x11="",
                           spectrum.save = c("sp0", "sp1"))
plot_double_spectrum_air <- 
     plot_double_spectrum(air_seas, series_name = "AirPassengers",
                          this_color = c("steelblue", "forestgreen"),  
                          this_median_color = c("blue", "green"), 
                          this_freq_color = c("darkblue", "darkgreen"), 
                          this_peak_color = c("red", "orange"))
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
