% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fcst.r
\name{plot_fcst}
\alias{plot_fcst}
\title{Forecast plot}
\usage{
plot_fcst(
  seas_obj = NULL,
  main_title = "ARIMA forecasts",
  sub_title = NULL,
  this_x_label = "Time",
  this_y_label = " ",
  length_ori = 2,
  do_grid = FALSE,
  do_background = FALSE,
  this_palette = c("darkgrey", "blue", "darkgreen", "darkgreen"),
  this_guide_legend = "Forecast"
)
}
\arguments{
\item{seas_obj}{\code{seas} object generated from a call of \code{seas} on a single time series
This is a required entry.}

\item{main_title}{Character string; main title of plot.
Default is \code{'ARIMA forecasts'}.}

\item{sub_title}{Subtitle for the plot.
Default is to generate the subtitle.}

\item{this_x_label}{Label for X-axis.  Default is \code{"Time"}}

\item{this_y_label}{Label for Y-axis.  Default is \code{" "}}

\item{length_ori}{Integer scalar; number of years of the original series
to show with forecasts.
Default is 2 years.}

\item{do_grid}{Logical scalar; indicates if certain plots will have grid lines.
Default is no grid lines.}

\item{do_background}{Logical scalar; indicates grey background included in plot.
Default is no grey background;}

\item{this_palette}{Array of character strings; color used for original series,
forecast, and upper and lower forecast bounds.
Default is \code{c("darkgrey", "blue", "darkgreen", "darkgreen")}.}

\item{this_guide_legend}{Title for legend.  Default is \code{"Forecast"}}
}
\value{
A \code{ggplot} object of the regARIMA forecasts with confidence bounds.
}
\description{
Generates a \code{ggplot} plot of regARIMA forecasts with confidence bounds.
}
\details{
Version 3.2, 11/6/2024
}
\examples{
air_seas <- seasonal::seas(AirPassengers, arima.model = "(0 1 1)(0 1 1)",
                           forecast.maxlead = 60, forecast.save = "fct",
                           series.save = "a1")
air_fcst_p <- 
    plot_fcst(air_seas, 
              main_title = "Forecasts for Airline Passengers", do_grid = TRUE)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
