% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fcst_history.r
\name{plot_fcst_history}
\alias{plot_fcst_history}
\title{Generate forecast history plot}
\usage{
plot_fcst_history(
  seas_mdl1 = NULL,
  seas_mdl2 = NULL,
  main_title = "Differences in the Sum of Squared Forecast Errors",
  name_mdl1 = "Model 1",
  name_mdl2 = "Model 2",
  do_grid = FALSE,
  do_background = FALSE,
  this_x_label = "Time",
  this_y_label = " ",
  this_palette = c("steelblue", "darkgreen"),
  this_guide_legend = "Fcst History"
)
}
\arguments{
\item{seas_mdl1}{\code{seas} object generated from a call of \code{seas}
on a single time series for the first model
This is a required entry.}

\item{seas_mdl2}{\code{seas} object generated from a call of \code{seas}
on a single time series for the second model
This is a required entry.}

\item{main_title}{Character string; main title of plot.
Default is \code{'Differences in the Sum of Squared Forecast Errors'}.}

\item{name_mdl1}{Character string; Description of first model for use in the subtitle.
Default is \code{'Model 1'}.}

\item{name_mdl2}{Character string; Description of second model for use in the subtitle.
Default is \code{'Model 2'}.}

\item{do_grid}{Logical scalar; indicates if certain plots will have grid lines.
Default is no grid lines.}

\item{do_background}{Logical scalar; indicates grey background included in plot.
Default is no grey background.}

\item{this_x_label}{Label for X-axis.  Default is \code{"Time"}}

\item{this_y_label}{Label for Y-axis.  Default is \code{" "}}

\item{this_palette}{Character array of length 2; color used for each forecast lag.
Default is \code{c("steelblue", "darkgreen")}.}

\item{this_guide_legend}{Title for legend.  Default is \code{"Fcst History"}}
}
\value{
A \code{ggplot} object that produces a forecast history plot.
If series not specified, print out error message and return NULL.
}
\description{
Generate forecast history plot, which compares the sum of squared forecast errors for two models.
}
\details{
Version 2.5 11/6/2024
}
\examples{
air_seas_mdl <- 
    seasonal::seas(AirPassengers, x11="", slidingspans = "", transform.function = "log",
                   arima.model = "(0 1 1)(0 1 1)", regression.aictest = NULL, outlier = NULL,
                   forecast.maxlead = 36, check.print = c( 'pacf', 'pacfplot' ),
                   history.fstep = c(1, 12), history.estimates = 'fcst', 
                   history.save = 'fcsterrors')
air_seas_mdl2 <- 
    seasonal::seas(AirPassengers, x11="", slidingspans = "", transform.function = "log",
                   arima.model = "(0 1 1)(0 1 1)", regression.variables = c("td"),
                   forecast.maxlead = 36, check.print = c( 'pacf', 'pacfplot' ),
                   history.fstep = c(1, 12), history.estimates = 'fcst', 
                   history.save = 'fcsterrors')
plot_fcst_history_air <-
    plot_fcst_history(air_seas_mdl, air_seas_mdl2, 
         main_title = 'Differences in the Sum of Squared Forecast Errors for Airline Passengers',
         name_mdl1 = 'Airline model', name_mdl2 = 'Airline model + regressors')
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
