% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_first_difference.r
\name{plot_first_difference}
\alias{plot_first_difference}
\title{First Difference Plot}
\usage{
plot_first_difference(
  this_series = NULL,
  take_log = FALSE,
  main_title = "First Difference Plot",
  remove_y_axis = TRUE,
  x_title_size = 10,
  geom_text_size = 2.5,
  geom_text_color = "steelblue"
)
}
\arguments{
\item{this_series}{\code{ts} object of a single time series
This is a required entry.}

\item{take_log}{Logical scalar, specifies that a log transformation will be
taken before differencing. Default is \code{FALSE}}

\item{main_title}{Title for the plot.
Default is character string \code{'First Difference Plot'}.}

\item{remove_y_axis}{Logical scalar.
If \code{TRUE}, removes the y-axis labels and tick marks from all subplots.
Default is \code{FALSE}, which keeps the y-axis labels and tick marks.}

\item{x_title_size}{Integer scalar, size of the x-axis title.
Default is \code{10}.}

\item{geom_text_size}{Integer scalar, size of the plotting characters.
Default is \code{2.5}.}

\item{geom_text_color}{Character scalar, color of the plotting characters.
Default is \code{"steelblue"}.}
}
\value{
A \code{ggplot} object that produces a plot of the first differences
of a series specified by \code{this_series} grouped by month or quarter.
The time series specified should be either a monthly or quarterly series.
}
\description{
Generates a ggplot2 object of the first difference of a time series grouped
by months or quarters.
}
\details{
Version 1.1, 8/22/2024
}
\examples{
shoes_seas <-
    seasonal::seas(shoes2008, x11.save = "d13",
       x11 = "", transform.function = "log",
       check.print = c("none", "+acf", "+acfplot", "+normalitytest"),
       regression.aictest = c("td", "easter"),
	  regression.save = c("td", "hol"),
       outlier.types = "all",
       arima.model = "(0 1 1)(0 1 1)",
       forecast.maxlead = 60)
shoesSA <- seasonal::final(shoes_seas)
p_shoes_sa_diff <- 
    plot_first_difference(shoesSA, 
        main_title = "Seasonal Adjustment First Differences",
        remove_y_axis = FALSE, x_title_size = 10)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
