% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fts.r
\name{plot_fts}
\alias{plot_fts}
\title{Final t-statistics for the outlier identification procedure plot}
\usage{
plot_fts(
  seas_obj = NULL,
  start_plot = NULL,
  main_title = "Outlier T-Values",
  this_y_label = NULL,
  this_x_label = "Time",
  add_identified_otl = FALSE,
  color_otl = NULL,
  this_palette = "Dark2"
)
}
\arguments{
\item{seas_obj}{\code{seas} object generated from a call of \code{seas} on a single time series
This is a requited entry.}

\item{start_plot}{Integer vector of length 2; Starting date for plot. Default is starting date for the time series.}

\item{main_title}{Character string; main title of plot.  Default is  \code{'Outlier T-Values'}.}

\item{this_y_label}{Character string; y-axis label for plot, if specified.}

\item{this_x_label}{Label for X axis. Default is \code{"Time"}.}

\item{add_identified_otl}{Logical scalar; indicates if outlier plots will include identified outliers.
Default is not including identified outliers.}

\item{color_otl}{Character array of length 3; color used for different outliers, with the order being \code{'ao','ls','tc'}.
Default is NULL.}

\item{this_palette}{Character string; default \code{RColorBrewer} palette.
Deault is \code{"Dark2"}.}
}
\value{
A \code{ggplot} object which produces a plot of the final t-statistics
from the automatic outlier identification procedure.
}
\description{
Generates a plot of the final t-statistics for the outlier identification procedure.
}
\details{
Version 2.6, 9/5/2024
}
\examples{
air_seas_outlier <- 
    seasonal::seas(AirPassengers, arima.model = "(0 1 1)(0 1 1)", 
                   outlier.save = "fts", outlier.types = "all")
plot_fts_air <- 
    plot_fts(air_seas_outlier, 
             main_title = "Outlier T-Values for Airline Passengers")
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{monsell.brian@gmail.com}
}
