% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fully_differenced_transformed.r
\name{plot_fully_differenced_transformed}
\alias{plot_fully_differenced_transformed}
\title{Plot fully differenced transformed series from an X-13ARIMA-SEATS SEATS
seasonal adjustment run.}
\usage{
plot_fully_differenced_transformed(
  seas_obj = NULL,
  this_series = "ori",
  main_title = NULL,
  sub_title = NULL,
  this_y_label = NULL,
  y_limit = NULL,
  this_x_label = "Time",
  start_plot = NULL,
  do_grid = FALSE,
  do_background = FALSE,
  draw_recess = FALSE,
  recess_color = "lightgrey",
  recess_sub = TRUE,
  line_color = "steelblue"
)
}
\arguments{
\item{seas_obj}{\code{seas} object generated from a call of \code{seas} on a single time series
This is a required entry.}

\item{this_series}{Character string; series for which SEATS produces a fully
differenced transformed series, limited to the original series (\code{ori},
the default), the seasonally adjusted series (\code{sa}), or the
trend component (\code{trend}). For other entries, the function will
print an error message and return a NULL.}

\item{main_title}{Character string; main title of plot.
A title will be generated if no title is specified.}

\item{sub_title}{Character string; subtitle of plot. There is no default subtitle.}

\item{this_y_label}{Character string; y-axis label for plot, if specified.}

\item{y_limit}{Numeric vector of length 2;
Range of values you wish the plot to be plotted over.
Default is range of the series specified.}

\item{this_x_label}{Label for X axis. Default is \code{"Time"}.}

\item{start_plot}{Integer vector of length 2; Starting date for plot.
Default is starting date for the time series.}

\item{do_grid}{Logical scalar; indicates if certain plots will have grid lines.
Default is no grid lines.}

\item{do_background}{Logical scalar; indicates grey background included in plot.
Default is no grey background;}

\item{draw_recess}{Logical scalar; indicates if certain plots will have
shaded areas for NBER recession dates.
Default is no recession shading.}

\item{recess_color}{Character string; color used for shading of recession region.
Default is \code{'lightgrey'}.}

\item{recess_sub}{Logical scalar; indicates if x-axis label for recession is
produced for this plot.
Default is x-axis label.}

\item{line_color}{Character string; color used for series in the plot.
Default is \code{'steelblue'}.}
}
\value{
A \code{ggplot} object that generates a plot of the fully differenced transformed series
from a SEATS seasonal adjustment.
If SEATS seasonal adjustment not produced, print out error message and
return NULL.
}
\description{
Generate plot of the fully differenced transformed series from a SEATS adjustment
from a \code{seas} object generated by the \code{seasonal} package.
}
\details{
Version 1.6, 8/29/2024
}
\examples{
shoes_seats_seas <- 
   seasonal::seas(shoes2008, arima.model = "(0 1 1)(0 1 1)", 
                  transform.function = "log", 
                  forecast.maxlead = 36,
                  check.print = c( 'pacf', 'pacfplot' ), 
                  seats.finite = "yes",
                  seats.save = c( 'dor', 'dsa', 'dtr' ) )
p_ori_full_diff <- plot_fully_differenced_transformed(shoes_seats_seas, "ori",
	sub_title = "US Shoe Sales",
     do_grid = TRUE, 
	line_color = "darkgreen")
p_sa_full_diff <- plot_fully_differenced_transformed(shoes_seats_seas, "sa",
	sub_title = "US Shoe Sales",
     do_grid = TRUE, 
	line_color = "darkgreen")
p_trn_full_diff <- plot_fully_differenced_transformed(shoes_seats_seas, "trn",
	sub_title = "US Shoe Sales",
     do_grid = TRUE, 
	line_color = "darkgreen")
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
