% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matrix.r
\name{plot_matrix}
\alias{plot_matrix}
\title{Plot time series matrix}
\usage{
plot_matrix(
  this_matrix = NULL,
  main_title = deparse(substitute(this_matrix)),
  sub_title = NULL,
  this_y_label = NULL,
  y_limit = NULL,
  this_x_label = "Time",
  start_plot = NULL,
  do_grid = FALSE,
  do_background = FALSE,
  line_color = NULL,
  this_palette = "Dark2",
  this_line_type = "solid",
  do_facet = FALSE,
  reset_facet_y_axis = FALSE
)
}
\arguments{
\item{this_matrix}{Numeric matrix; columns of time series object to be plotted.}

\item{main_title}{Character string; main title of plot.
The default title is the name of the matrix passed to this function.}

\item{sub_title}{Character string; subtitle of plot. There is no default subtitle.}

\item{this_y_label}{Character string; y-axis label for plot, if specified.}

\item{y_limit}{Numeric vector of length 2; Range of values on plot y-axis
Default is range of the series specified.}

\item{this_x_label}{Label for X axis. Default is \code{"Time"}.}

\item{start_plot}{Integer vector of length 2; Starting date for plot.
Default is starting date for the time series.}

\item{do_grid}{Logical scalar; indicates if plots will have grid lines.
Default is no grid lines.}

\item{do_background}{Logical scalar; indicates grey background included in plot.
Default is no grey background;}

\item{line_color}{Character scalar; color used for plot.
User should specify one color for each column of the matrix specified.
Default is the \code{RColorBrewer} palette \code{"Dark2"}.}

\item{this_palette}{Character string; default \code{RColorBrewer} palette.
Deault is \code{"Dark2"}.}

\item{this_line_type}{Character string; indicates line type of each plot produced.
Default is \code{"solid"}.}

\item{do_facet}{Logical scalar; indicates if a facet plot is generated of the
different colums.
Default is \code{FALSE}.}

\item{reset_facet_y_axis}{Logical scalar; indicates if y-axis for facet plots
reset to \code{y_limit}
Default is \code{FALSE}.}
}
\value{
A \code{ggplot} object that produces a plot of user-specified time series.
If matrix not specified, print out error message and return NULL.
}
\description{
Generate plot of a matrix of user-specified time series.
}
\details{
Version 4.1, 9/5/2024
}
\examples{
BP_Region_Matrix <-
   cbind(blsplotGG::xt_data_list$mw1u, blsplotGG::xt_data_list$ne1u, 
         blsplotGG::xt_data_list$so1u, blsplotGG::xt_data_list$we1u)
colnames(BP_Region_Matrix) <- names(blsplotGG::xt_data_list)
p_BP <- blsplotGG::plot_matrix(BP_Region_Matrix, this_y_label = 'Building Permits', 
   main_title = "US Building Permits, 1 Family Units",
   do_grid = TRUE, this_line_type = rep("solid", 4),
   line_color = c("orange", "steelblue", "forestgreen", "brown"))
p_BP_facet <- blsplotGG::plot_matrix(BP_Region_Matrix, this_y_label = 'Building Permits', 
   main_title = "US Building Permits, 1 Family Units",
   do_grid = FALSE, do_facet = TRUE,
   line_color = c("orange", "steelblue", "forestgreen", "brown"))
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
