% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maximum_percent_difference.r
\name{plot_maximum_percent_difference}
\alias{plot_maximum_percent_difference}
\title{Maximum percent difference plot}
\usage{
plot_maximum_percent_difference(
  seas_obj = NULL,
  this_series = "sfs",
  main_title = "Maximum Percent Difference Plot",
  sub_title = NULL,
  this_x_label = "Time",
  this_y_label = " ",
  do_grid = FALSE,
  do_background = FALSE,
  line_color = "steelblue",
  cut_color = "red"
)
}
\arguments{
\item{seas_obj}{\code{seas} object generated from a call of \code{seas} on a single time series
This is a required entry.}

\item{this_series}{Character string; three character code for the type of series to be generated.
Allowed entries are \code{"sfs"} (seasonal factors, default),
\code{"chs"} (period-to-period changes),
\code{"sis"} (indirect seasonal factors),
\code{"cis"} (indirect period-to-period changes).}

\item{main_title}{Title for the plot.
Default is character string \code{'Maximum Percent Difference Plot'}.}

\item{sub_title}{Subtitle for the plot. Optional entry.}

\item{this_x_label}{Label for X-axis.  Default is \code{"Time"}}

\item{this_y_label}{Label for Y-axis.  Default is \code{" "}}

\item{do_grid}{Logical scalar; indicates if certain plots will have grid lines.
Default is no grid lines.}

\item{do_background}{Logical scalar; indicates grey background included in plot.
Default is no grey background;}

\item{line_color}{Color used for lines in the maximum percentage difference plot.
Default is \code{"steelblue"}.}

\item{cut_color}{Color use to show the slidings spans cut off for this type of series.
Default is \code{"red"}.}
}
\value{
A \code{ggplot} object that generates a plot of the maximum percent difference
from a sliding spans analysis of seasonal factors or changes.
}
\description{
Generates a \code{ggplot} object with a time series of the maximum percent difference from
a sliding spans analysis of seasonal factors or changes.
}
\details{
Version 1.4, 11/7/2024
}
\examples{
shoes_sspan_seas <- 
   seasonal::seas(blsplotGG::shoes2008, 
                  arima.model = "(0 1 1)(0 1 1)", 
                  transform.function = "log", 
                  forecast.maxlead = 36,
                  slidingspans.save = c('sfs', 'chs'))
p_shoes_maxpct_sf <- 
    plot_maximum_percent_difference(shoes_sspan_seas, "sfs",
                       main_title = "Maximum Percent Difference Plot of Seasonal Factors",
                       sub_title = "US Shoe Sales",
                       line_color = "blue", cut_color = "purple")
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
