% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maximum_percent_difference_by_period.r
\name{plot_maximum_percent_difference_by_period}
\alias{plot_maximum_percent_difference_by_period}
\title{Maximum percent difference for sliding spans by period}
\usage{
plot_maximum_percent_difference_by_period(
  seas_obj = NULL,
  this_series = "sfs",
  main_title = NULL,
  sub_title = NULL,
  this_y_label = NULL,
  this_x_label = NULL,
  do_grid = FALSE,
  do_background = FALSE,
  line_color = NULL,
  this_palette = "Set3",
  cut_color = "red"
)
}
\arguments{
\item{seas_obj}{\code{seas} object generated from a call of \code{seas} on a single time series
This is a required entry.}

\item{this_series}{Character string; three character code for the type of series to be generated.
Allowed entries are \code{"sfs"} (seasonal factors, default),
\code{"chs"} (period-to-period changes),
\code{"sis"} (indirect seasonal factors),
\code{"cis"} (indirect period-to-period changes).}

\item{main_title}{Character string; main title of plot.
Default is character string \code{'Maximum Percent Difference Plot by Period'},
where Period is replaced by Month or Quarter.}

\item{sub_title}{Character string; subtitle of plot. There is no default subtitle.}

\item{this_y_label}{Character string; y-axis label for plot, if specified.}

\item{this_x_label}{Label for X axis. Default is \code{"Month"} or \code{"Quarter"}.}

\item{do_grid}{Logical scalar; indicates if plots will have grid lines.
Default is no grid lines.}

\item{do_background}{Logical scalar; indicates grey background included in plot.
Default is no grey background.}

\item{line_color}{Character scalar; color used for plot.
User should specify one color for each column of the matrix specified.
Default is the \code{RColorBrewer} palette \code{"Set3"}.}

\item{this_palette}{Character string; default \code{RColorBrewer} palette.
Default is \code{"Set3"}.}

\item{cut_color}{Color use to show the slidings spans cut off for this type of series.
Default is \code{"red"}.}
}
\value{
Generate \code{ggplot} object generating a plot of the maximum percent
difference from a sliding spans analysis by month or quarter
}
\description{
Generate a plot of the maximum percent difference from a sliding spans analysis by
month or quarter.
}
\details{
Version 1.4, 11/7/2024
}
\examples{
shoes_sspan_seas <- 
   seasonal::seas(blsplotGG::shoes2008, 
                  arima.model = "(0 1 1)(0 1 1)", 
                  transform.function = "log", 
                  forecast.maxlead = 36,
                  slidingspans.save = c('sfs', 'chs'))
p_shoes_maxpct_sf <- 
    plot_maximum_percent_difference_by_period(shoes_sspan_seas, "sfs",
                       main_title = "Maximum Percent Difference Plot of Seasonal Factors",
                       sub_title = "US Shoe Sales",
                       this_palette = "Paired")
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
