% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ratio.r
\name{plot_ratio}
\alias{plot_ratio}
\title{Ratio plot}
\usage{
plot_ratio(
  ratio_series = NULL,
  ratio_range = NULL,
  main_title = "Ratio Plot",
  sub_title = NULL,
  this_x_label = "Time",
  this_y_label = "Ratio",
  do_grid = FALSE,
  do_background = FALSE,
  draw_recess = FALSE,
  recess_color = "lightgrey",
  recess_sub = TRUE,
  add_line = TRUE,
  ratio_mean = 1,
  ratio_color = "steelblue"
)
}
\arguments{
\item{ratio_series}{Time series of ratios/factors for which you want to generate
a high definition plot}

\item{ratio_range}{Range of values you wish the plot to be plotted over.
Default is range of the series.}

\item{main_title}{Title for the plot.
Default is character string \code{'Ratio Plot'}.}

\item{sub_title}{Subtitle for the plot. Default is \code{NULL}.}

\item{this_x_label}{Label for X axis. Default is \code{"Time"}.}

\item{this_y_label}{Label for Y axis. Default is \code{"Ratio"}.}

\item{do_grid}{Logical scalar; indicates if certain plots will have grid lines.
Default is no grid lines.}

\item{do_background}{Logical scalar; indicates grey background included in plot.
Default is no grey background;}

\item{draw_recess}{Logical scalar; indicates if certain plots will have shaded
areas for NBER recession dates.
Default is no recession shading.}

\item{recess_color}{Character string; color used for shading of recession region.
Default is \code{'lightgrey'}.}

\item{recess_sub}{Logical scalar; indicates if x-axis label for recession is
produced for this plot.
Default is x-axis label.}

\item{add_line}{Logical scalar; add solid line for assumed mean.
Default is \code{TRUE}.}

\item{ratio_mean}{Assumed mean value for the ratio.  Default is \code{1.0}}

\item{ratio_color}{Color used for lines in ratio plot.
Default is \code{"steelblue"}.}
}
\value{
A \code{ggplot} object that produces a ratio plot of a time series.
}
\description{
Generates a high-definition plot around a reference line other than zero.
}
\details{
Version 3.1, 8/28/2024
}
\examples{
ukgas_x11_seas <- 
   seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)", 
                  transform.function = "log", forecast.maxlead = 20,
                  x11.seasonalma = "s3x5", check.print = c( 'pacf', 'pacfplot' ))
ukgas_x11_sf     <- seasonal::series(ukgas_x11_seas, "d10")
ukgas_sf_ratio   <- plot_ratio(ukgas_x11_sf, main_title = "UK Gas X-11 Seasonal Factors",
                               sub_title = "Seasonal Filter: 3x5")
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
