% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ratio_facet.r
\name{plot_ratio_facet}
\alias{plot_ratio_facet}
\title{Ratio facet plot}
\usage{
plot_ratio_facet(
  seas_obj = NULL,
  ratio_tables = NULL,
  main_title = "Ratio Facet Plot",
  sub_title = NULL,
  this_x_label = "Time",
  this_y_label = "Ratio",
  do_grid = FALSE,
  do_background = FALSE,
  add_line = TRUE,
  ratio_mean = 1,
  ratio_color = "steelblue"
)
}
\arguments{
\item{seas_obj}{\code{seas} object generated from a call of \code{seas}
on a single time series
This is a required entry.}

\item{ratio_tables}{Array of tables for which you want to generate a
high definition plot. Possible entries are \code{"sf"} (seasonal factor),
\code{"td"} (trading day factor), \code{"hol"} (holiday factors), and
\code{"irr"} (irregular).}

\item{main_title}{Title for the plot.
Default is character string \code{'Ratio Facet Plot'}.}

\item{sub_title}{Subtitle for the plot. Default is \code{NULL}.}

\item{this_x_label}{Label for X axis. Default is \code{"Time"}.}

\item{this_y_label}{Label for Y axis. Default is \code{"Ratio"}.}

\item{do_grid}{Logical scalar; indicates if certain plots will have grid lines.
Default is no grid lines.}

\item{do_background}{Logical scalar; indicates grey background included in plot.
Default is no grey background;}

\item{add_line}{Logical scalar; add solid line for assumed mean.
Default is \code{TRUE}.}

\item{ratio_mean}{Assumed mean value for the ratio.  Default is \code{1.0}}

\item{ratio_color}{Color used for lines in ratio plot.
Default is \code{"steelblue"}.}
}
\value{
A \code{ggplot} object that generates a ratio facet plot for the
factors provided in the \code{ratio_tables} argument.
}
\description{
Generates a high-definition plot for a number of X-13 factors around
a reference line other than zero.
}
\details{
Version 2.5, 8/28/2024
}
\examples{
shoes_x11_seas <- 
   seasonal::seas(shoes2008, arima.model = "(0 1 1)(0 1 1)", 
                  transform.function = "log", forecast.maxlead = 36,
                  regression.aictest = c("td", "easter"),
	             regression.save = c("td", "hol"),
                  x11.seasonalma = "s3x5", check.print = c( 'pacf', 'pacfplot' ))
shoes_factor_ratio_one   <- 
   plot_ratio_facet(shoes_x11_seas, c("sf", "irr"), 
                    main_title = "US Retail Shoe Sales",
                    sub_title = "X-11 Seasonal (3x5) and Irregular Factors")
shoes_factor_ratio_two   <- 
   plot_ratio_facet(shoes_x11_seas, c("td", "hol", "irr"), 
                    main_title = "US Retail Shoe Sales",
                    sub_title = "Trading Day, Easter, and Irregular Factors")
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{monsell.brian@gmail.com}
}
