% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ratio_two.r
\name{plot_ratio_two}
\alias{plot_ratio_two}
\title{Ratio plots for two series}
\usage{
plot_ratio_two(
  ratio_one = NULL,
  ratio_two = NULL,
  ratio_range = NULL,
  do_comparison_plot = TRUE,
  overall_title = NULL,
  ratio_one_title = "First Ratio Plot",
  ratio_two_title = "Second Ratio Plot",
  comparison_title = "Comparison Plot",
  this_x_label = NULL,
  this_y_label = NULL,
  do_grid = FALSE,
  do_background = FALSE,
  add_line = TRUE,
  ratio_mean = c(1, 1),
  ratio_color = "steelblue"
)
}
\arguments{
\item{ratio_one}{First time series of ratios/factors for which you want to generate
a high definition plot.
This is a required entry.}

\item{ratio_two}{Second time series of ratios/factors for which you want to generate
a high definition plot.
This is a required entry.}

\item{ratio_range}{Range of values you wish the plot to be plotted over.
Default is range of the series, if they are the same type of factor.}

\item{do_comparison_plot}{Logical scalar. If TRUE, a ratio/difference plot of the two
factors will be generated from \code{ratio_one} and \code{ratio_two},
if the series have the same periodicity and are the same type of factor.
Default is TRUE.}

\item{overall_title}{Title for the combined plot.
Default is NULL.}

\item{ratio_one_title}{Title for the first plot.
Default is character string \code{'First Ratio Plot'}.}

\item{ratio_two_title}{Title for the plot.
Default is character string \code{'Second Ratio Plot'}.}

\item{comparison_title}{Title for the comparison plot, if requested.
Default is character string \code{'Comparison Plot'}.}

\item{this_x_label}{Label for X axis. Default is no X axis label.}

\item{this_y_label}{Label for Y axis. Default is no Y axis label.}

\item{do_grid}{Logical scalar; indicates if certain plots will have grid lines.
Default is no grid lines.}

\item{do_background}{Logical scalar; indicates grey background included in plot.
Default is no grey background;}

\item{add_line}{Logical scalar; add solid line for assumed mean.
Default is \code{TRUE}.}

\item{ratio_mean}{Numeric vector of length two; the assumed mean value for the
ratio of each series.  Default is \code{c(1.0, 1.0)}}

\item{ratio_color}{Color used for lines in ratio plot.
Default is \code{"steelblue"}.}
}
\value{
A \code{ggplot} object that generates a stacked plot with the
}
\description{
Generates ratio plots for two series, and a third plot of the ratio/difference of
the two series if the user requests it.
}
\details{
Version 1.4, 9/6/2024
}
\examples{
ukgas_x11_seas   <- 
   seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)", 
                  transform.function = "log", forecast.maxlead = 20,
                  x11.seasonalma = "s3x5", x11.save = c("d10", "d11"),
                  check.print = c( 'pacf', 'pacfplot' ))
ukgas_seats_seas <- 
   seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)", 
                  transform.function = "log", forecast.maxlead = 20,
                  seats.save = c("s10", "s11"),
                  check.print = c( 'pacf', 'pacfplot' ))
ukgas_x11_sf      <- seasonal::series(ukgas_x11_seas, "d10")
ukgas_seats_sf    <- seasonal::series(ukgas_seats_seas, "s10")
ukgas_sf_two_plot <- 
     plot_ratio_two(ukgas_x11_sf, ukgas_seats_sf,
         overall_title = "UK Gas Production",
		ratio_one_title = "X-11 Seasonal Factors",
		ratio_two_title = "SEATS Seasonal Factors",
		comparison_title = "Ratio of Seasonal Factors (X11 / SEATS)",
         ratio_color = "darkgreen")
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
