% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_resid.r
\name{plot_resid}
\alias{plot_resid}
\title{Residual plot}
\usage{
plot_resid(
  seas_obj = NULL,
  main_title = "ARIMA Residuals",
  series_name = NULL,
  this_x_label = "Time",
  this_y_label = NULL,
  do_grid = TRUE,
  do_background = FALSE,
  draw_recess = FALSE,
  recess_color = NULL,
  recess_sub = TRUE,
  use_ratio = FALSE,
  add_line = TRUE,
  line_color = "steelblue"
)
}
\arguments{
\item{seas_obj}{\code{seas} object generated from a call of \code{seas} on a single time series
This is a requited entry.}

\item{main_title}{Character string; main title of plot.
Default is  \code{'ARIMA Residuals'}.}

\item{series_name}{Character scalar; name of the time series used in \code{seas_obj}.}

\item{this_x_label}{Label for X axis. Default is \code{"Time"}.}

\item{this_y_label}{Label for Y axis. Default is \code{series_name}. if specified.}

\item{do_grid}{Logical scalar; indicates if certain plots will have grid lines.
Default is grid lines plotted.}

\item{do_background}{Logical scalar; indicates grey background included in plot.
Default is no grey background;}

\item{draw_recess}{Logical scalar; indicates if certain plots will have shaded
areas for NBER recession dates. Default is recession shading not plotted.}

\item{recess_color}{Character string; color used for shading of recession region.
Default is \code{'lightgrey'}.}

\item{recess_sub}{Logical scalar; indicates if x-axis label for recession is
produced for this plot. Default is x-axis label is produced}

\item{use_ratio}{Logical scalar; indicates if plots of seasonal factors, irregular,
and residuals are done as ratio plots.
Default has these plots as time series line plots.}

\item{add_line}{Logical scalar; add solid line for assumed mean.
Default is \code{TRUE}.}

\item{line_color}{Character string; color used for residuals.
Default is \code{"green"}.}
}
\value{
Generates a \code{ggplot} object of a plot of the regARIMA residuals with
diagnostic information in the sub-headers.
}
\description{
Generates a plot of the regARIMA residuals with diagnostic information
}
\details{
Version 2.5, 8/26/2024
}
\examples{
air_seas <- seasonal::seas(AirPassengers, arima.model = "(0 1 1)(0 1 1)")
plot_resid_air <-
     plot_resid(air_seas, main_title = "ARIMA Residuals for Airline Passengers", 
                use_ratio = TRUE, line_color="darkblue")
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{monsell.brian@gmail.com}
}
