% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sa_history.r
\name{plot_sa_history}
\alias{plot_sa_history}
\title{Revisions History Plot for Seasonal Adjustments}
\usage{
plot_sa_history(
  seas_obj = NULL,
  add_ori = TRUE,
  main_title = "Seasonal Adjustment History Graph",
  sub_title = NULL,
  this_x_label = "Time",
  this_y_label = NULL,
  do_grid = FALSE,
  do_background = FALSE,
  line_color = NULL,
  this_palette = "Dark2"
)
}
\arguments{
\item{seas_obj}{\code{seas} object generated from a call of \code{seas} on a single time series
This is a required entry.}

\item{add_ori}{Logical scalar; add the original series to the plot. Default is \code{TRUE}.}

\item{main_title}{Character string. Title for the plot.
Default is \code{'Seasonal Adjustment History Graph'}.}

\item{sub_title}{Subtitle for the plot. Default is \code{NULL}.}

\item{this_x_label}{Label for X axis. Default is \code{"Time"}.}

\item{this_y_label}{Label for Y axis. Default is \code{NULL}.}

\item{do_grid}{Logical scalar; indicates if certain plots will have grid lines.
Default is no grid lines.}

\item{do_background}{Logical scalar; indicates grey background included in plot.
Default is no grey background;}

\item{line_color}{Vector with colors used for lines in history plot.
Should be of length 4 (if \code{add_ori = FALSE}) or 5 (if \code{add_ori = TRUE})
Default is \code{NULL}, which indicates that the palette specified
in \code{this_palette} is used to generate colors for this plot.}

\item{this_palette}{Color used for lines in plot. Default is \code{"Dark2"}}
}
\value{
A \code{ggplot} object that generates a history plot of the seasonal adjustment.
The seas object in the \code{seas_obj} argument must contain output for a
revisions history analysis for seasonal adjustments with \code{sadjlags}
set to 1 and 12 (for monthly series) or 4 (for quarterly series).
}
\description{
Generates a ggplot2 object of estimates from a revisions history of a seasonal
adjustment for a given series.
}
\details{
Version 1.9, 11/6/2024
}
\examples{
shoes_seas <-
    seasonal::seas(shoes2008, x11.save = "d13",
       x11 = "", transform.function = "log",
       check.print = c("none", "+acf", "+acfplot", "+normalitytest"),
       regression.aictest = c("td", "easter"),
	  regression.save = c("td", "hol"),
       outlier.types = "all",
       arima.model = "(0 1 1)(0 1 1)",
       forecast.maxlead = 60,
	   	  history.estimates = c("sadj", "sadjchng", "trend"),
	      history.sadjlags = c(1,12),
	      history.print = "all",
	      history.save = c("sae", "sar"))
p_shoes_sa_history <- 
    plot_sa_history(shoes_seas, add_ori = FALSE, 
        main_title = "SA History Graph, Lag 1 and 12",
        sub_title = "US Retail Sales of Shoes")
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
