% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sadj_and_ori.r
\name{plot_sadj_and_ori}
\alias{plot_sadj_and_ori}
\title{Plot X-13 seasonal adjustment, original series on same axis}
\usage{
plot_sadj_and_ori(
  this_sa = NULL,
  this_ori = NULL,
  this_sa_type = "SEATS",
  main_title = NULL,
  sub_title = NULL,
  this_x_label = "Time",
  this_y_label = " ",
  do_grid = FALSE,
  do_background = FALSE,
  line_color = c("grey", "darkblue"),
  this_guide_legend = "Series"
)
}
\arguments{
\item{this_sa}{Time series of a seasonal adjustment.
This is a required entry.}

\item{this_ori}{Time series of the original series. Optional entry.}

\item{this_sa_type}{Character string; type of seasonal adjustment.
Default is \code{"SEATS"}.}

\item{main_title}{Title for the plot.
By default, the routine will generate a title based on the
type of adjustment (X-11 and SEATS) done.}

\item{sub_title}{Subtitle for the plot. Optional entry.}

\item{this_x_label}{Label for X-axis.  Default is \code{"Time"}}

\item{this_y_label}{Label for Y-axis.  Default is \code{" "}}

\item{do_grid}{Logical scalar; indicates if certain plots will have grid lines.
Default is no grid lines.}

\item{do_background}{Logical scalar; indicates grey background included in plot.
Default is no grey background;}

\item{line_color}{Character vector of length 2; color used for lines in the plot,
in the order of original series, seasonally adjusted series.
Default is \code{c("grey", "darkblue")}.}

\item{this_guide_legend}{Title for legend.  Default is \code{"Series"}}
}
\value{
A \code{ggplot} object that generates a plot comparing a
seasonally adjusted series with the original series.
}
\description{
Generates a \code{ggplot} object with a time series plot that compares an X-13
seasonal adjustment and trend, optionally including the original series.
}
\details{
Version 2.2, 12/17/2024
}
\examples{
ukgas_x11_seas   <- 
   seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)", 
                  transform.function = "log", forecast.maxlead = 20,
                  x11.seasonalma = "s3x5", 
                  check.print = c( 'pacf', 'pacfplot' ))
ukgas_x11_sa     <- seasonal::final(ukgas_x11_seas)
ukgas_sa_ori_p <- 
    plot_sadj_and_ori(this_sa = ukgas_x11_sa, 
                      this_ori = UKgas,
					 this_sa_type = "X-11",
                      main_title = "UK Gas Seasonal Adjustment",
                      line_color = c("lightgrey", "forestgreen"))
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
