% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seasonal_sums.r
\name{plot_seasonal_sums}
\alias{plot_seasonal_sums}
\title{Plot of the seasonal period length sums of the SEATS seasonal factors from
an X-13ARIMA-SEATS SEATS seasonal adjustment run.}
\usage{
plot_seasonal_sums(
  seas_obj = NULL,
  main_title = NULL,
  sub_title = NULL,
  this_y_label = NULL,
  this_x_label = "Time",
  do_grid = FALSE,
  do_background = FALSE,
  draw_recess = FALSE,
  recess_color = "lightgrey",
  recess_sub = TRUE,
  line_color = "steelblue"
)
}
\arguments{
\item{seas_obj}{\code{seas} object generated from a call of \code{seas} on a single time series
This is a required entry.}

\item{main_title}{Character string; main title of plot.
A title will be generated if no title is specified.}

\item{sub_title}{Character string; subtitle of plot. There is no default subtitle.}

\item{this_y_label}{Character string; y-axis label for plot, if specified.}

\item{this_x_label}{Label for X axis. Default is \code{"Time"}.}

\item{do_grid}{Logical scalar; indicates if certain plots will have grid lines.
Default is no grid lines.}

\item{do_background}{Logical scalar; indicates grey background included in plot.
Default is no grey background;}

\item{draw_recess}{Logical scalar; indicates if certain plots will have
shaded areas for NBER recession dates.
Default is no recession shading.}

\item{recess_color}{Character string; color used for shading of recession region.
Default is \code{'lightgrey'}.}

\item{recess_sub}{Logical scalar; indicates if x-axis label for recession is
produced for this plot.
Default is x-axis label.}

\item{line_color}{Character string; color used for series in the plot.
Default is \code{'steelblue'}.}
}
\value{
A \code{ggplot} object which generates a plot of the seasonal period length sums
of the SEATS seasonal factors.
If SEATS seaonal adjustent not producted, print out error message and
return NULL.
}
\description{
Generate plot of the seasonal period length sums of the SEATS seasonal factors
from a SEATS adjustment from a \code{seas} object generated by the \code{seasonal}
package.
}
\details{
Version 1.5, 8/29/2024
}
\examples{
shoes_seats_seas <- 
   seasonal::seas(shoes2008, arima.model = "(0 1 1)(0 1 1)", 
                  transform.function = "log", 
                  forecast.maxlead = 36,
                  check.print = c( 'pacf', 'pacfplot' ), 
                  seats.finite = "yes",
                  seats.save = c( 'ssm' ) )
p_seasonal_sum <- plot_seasonal_sums(shoes_seats_seas, 
	sub_title = "US Shoe Sales",
     do_grid = TRUE, 
	line_color = "darkgreen")
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{monsell.brian@gmail.com}
}
