% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seats_filter.r
\name{plot_seats_filter}
\alias{plot_seats_filter}
\title{Plot adjustment or trend filter from an X-13ARIMA-SEATS SEATS seasonal
adjustment run.}
\usage{
plot_seats_filter(
  seas_obj = NULL,
  this_series = "sa",
  this_filter = "symmetric",
  main_title = NULL,
  sub_title = NULL,
  this_y_label = "Filter Coefficient",
  this_x_label = "Index",
  do_grid = FALSE,
  do_background = FALSE,
  point_color = "steelblue",
  point_size = 1.5,
  point_shape = 20
)
}
\arguments{
\item{seas_obj}{\code{seas} object generated from a call of \code{seas} on a single time series
This is a required entry.}

\item{this_series}{Character string; series for which SEATS produces a set of
filter weights, limited to the seasonally adjusted series
(\code{sa}, the default), or the trend component (\code{trend}).
For other entries, the function will print an error message and return a NULL.}

\item{this_filter}{Character string; type of filter for which SEATS produces
filter coefficients, limited to the symmetric (\code{symmetric},
the default), or the concurrent (\code{trend}) filter.
For other entries, the function will print an error message and return a NULL.}

\item{main_title}{Character string; main title of plot.
A title will be generated if no title is specified.}

\item{sub_title}{Character string; subtitle of plot. There is no default subtitle.}

\item{this_y_label}{Character string; y-axis label for plot.
Default is \code{"Filter Coefficient"}.}

\item{this_x_label}{Label for X axis. Default is \code{"Index"}.}

\item{do_grid}{Logical scalar; indicates if certain plots will have grid lines.
Default is no grid lines.}

\item{do_background}{Logical scalar; indicates grey background included in plot.
Default is no grey background;}

\item{point_color}{Character string; color used for points in the plot.
Default is \code{'steelblue'}.}

\item{point_size}{Integer scalar; relative size of points in filter plot.
Default is \code{1.5}.}

\item{point_shape}{Integer scalar; code for the shape of points in filter plot.
Default is \code{20}, a small filled circle.}
}
\value{
A \code{ggplot} object that generates a plot of the adjustment or trend filter
from a SEATS seasonal adjustment from X-13ARIMA-SEATS.
If SEATS seaonal adjustment (with \code{finite = yes}) not producted,
print out error message and return NULL.
}
\description{
Generate plot of the fully differenced transformed series from a SEATS adjustment
from a \code{seas} object generated by the \code{seasonal} package.
}
\details{
Version 2.2, 8/29/2024
}
\examples{
shoes_seats_seas <- 
   seasonal::seas(shoes2008, arima.model = "(0 1 1)(0 1 1)", 
                  transform.function = "log", 
                  forecast.maxlead = 36,
                  check.print = c( 'pacf', 'pacfplot' ), 
                  seats.finite = "yes",
                  seats.save = c( 'fac', 'faf', 'ftc', 'ftf' ) )
p_sa_sym_filter <- plot_seats_filter(shoes_seats_seas, "sa", "symmetric",
	sub_title = "US Shoe Sales",
	point_color = "darkgreen")
p_trn_sym_filter <- plot_seats_filter(shoes_seats_seas, "trn", "symmetric",
	sub_title = "US Shoe Sales",
	point_color = "darkgreen")
p_sa_conc_filter <- plot_seats_filter(shoes_seats_seas, "sa", "concurrent",
	sub_title = "US Shoe Sales",
	point_color = "darkgreen")
p_trn_conc_filter <- plot_seats_filter(shoes_seats_seas, "trn", "concurrent",
	sub_title = "US Shoe Sales",
	point_color = "darkgreen")
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{monsell.brian@gmail.com}
}
