% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sf_mean.r
\name{plot_sf_mean}
\alias{plot_sf_mean}
\title{Seasonal factor mean plot using ggplot}
\usage{
plot_sf_mean(
  this_sf_matrix = NULL,
  main_title = deparse(substitute(this_sf_matrix)),
  sub_title = NULL,
  this_y_label = NULL,
  this_x_label = "Time",
  do_grid = FALSE,
  do_background = FALSE,
  line_color = NULL,
  this_palette = "Dark2",
  this_line_type = rep("solid", ncol(this_sf_matrix)),
  do_facet = FALSE,
  reset_facet_y_axis = FALSE,
  y_limit = NULL,
  forecast = 0,
  this_legend_title = "SF Means",
  this_legend_entry = colnames(this_sf_matrix)
)
}
\arguments{
\item{this_sf_matrix}{time series object of the seasonal factors from a seasonal adjustment}

\item{main_title}{Character string; main title of plot.  Default is  \code{'Mean of Seasonal Factors'}.}

\item{sub_title}{Character string; subtitle of plot. There is no default subtitle.}

\item{this_y_label}{Character string; y-axis label for plot, if specified.}

\item{this_x_label}{Label for X axis. Default is \code{"Time"}.}

\item{do_grid}{Logical scalar; indicates if plots will have grid lines.
Default is no grid lines.}

\item{do_background}{Logical scalar; indicates grey background included in plot.
Default is no grey background;}

\item{line_color}{Character scalar; color used for plot. User should specify one
color for each column of the matrix specified.
Default is the \code{RColorBrewer} palette \code{"Dark2"}.}

\item{this_palette}{Character string; default \code{RColorBrewer} palette
Deault is \code{"Dark2"}.}

\item{this_line_type}{Character string; indicates line type of each plot produced.
Default is \code{rep("solid", ncol(this_sf_matrix))}.}

\item{do_facet}{Logical scalar; indicates if a facet plot is generated of the different colums.
Default is \code{FALSE}.}

\item{reset_facet_y_axis}{Logical scalar; indicates if y-axis for facet plots reset to \code{y_limit}
Default is \code{FALSE}.}

\item{y_limit}{Numeric vector of length 2; Range of values you wish the plot to be plotted over.
Default is range of the seasonal factors.}

\item{forecast}{Integer scalar; Number of forecasts appended to the seasonal factors. Default is 0.}

\item{this_legend_title}{Character string; indicates title of legend. Default is \code{'Series'}.}

\item{this_legend_entry}{Character array; entries for the legend. When \code{do_facet = TRUE},
entries are used as the facet labels.
Default is \code{colnames(this_sf_matrix)}}
}
\value{
Generate plot of the means of seasonal factors by period.
If seasonal factors not specified, print out error message and return NULL.
}
\description{
Generates a plot of the means of the seasonal factors
}
\details{
Version 2.7, 9/5/2024
}
\examples{
EM_individual_seas <-
 seasonal::seas(
    x11 = "", transform.function = "log",
    check.print = c("none", "+acf", "+acfplot", "+normalitytest"),
    regression.aictest = NULL,
    outlier.types = "all",
    arima.model = "(0 1 1)(0 1 1)",
    forecast.maxlead = 60,
    list = list(
        list(x = employment_list$n2000013),
        list(x = employment_list$n2000014),
        list(x = employment_list$n2000025),
        list(x = employment_list$n2000026)
    )
)
EM_names <- names(employment_list)
# Use Filter function to grab seas objects
EM_individual_seas_only <-
  Filter(function(x) inherits(x, "seas"), EM_individual_seas)

names(EM_individual_seas_only) <- EM_names

EM_Comp_Sf <-
 cbind(seasonal::series(EM_individual_seas_only$n2000013, "d10"),
      seasonal::series(EM_individual_seas_only$n2000014, "d10"),
       seasonal::series(EM_individual_seas_only$n2000025, "d10"),
       seasonal::series(EM_individual_seas_only$n2000026, "d10"))
colnames(EM_Comp_Sf) <- EM_names

em_plot <- blsplotGG::plot_sf_mean(EM_Comp_Sf,
  main_title = 'US Employment Seasonal Means',
  sub_title = 'X-11 Seasonals',
  forecast = 60, 
  this_legend_title = "SF Means",
  this_legend_entry = c("M 16-19", "F 16-19", "M 20+", "F 20+")
  )

em_plot_facet <- blsplotGG::plot_sf_mean(EM_Comp_Sf,
  main_title = 'US Employment Seasonal Means',
  sub_title = 'X-11 Seasonals',
  forecast = 60, 
  do_facet = TRUE, 
  reset_facet_y_axis = TRUE,
  this_legend_entry = c("M 16-19", "F 16-19", "M 20+", "F 20+")
  )
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{monsell.brian@gmail.com}
}
