% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sf_series.r
\name{plot_sf_series}
\alias{plot_sf_series}
\title{Seasonal factor plot grouped by month/quarter}
\usage{
plot_sf_series(
  this_sf = NULL,
  y_limit = NULL,
  this_trans = TRUE,
  main_title = "Seasonal Sub-Plots",
  sub_title = NULL,
  this_xlab = NULL,
  do_grid = FALSE,
  do_background = FALSE,
  this_color_sf = "darkblue",
  this_color_mean = "darkgrey",
  first_year = NULL,
  add_mean_line = TRUE,
  this_legend_title = "SF Plot",
  this_legend_text = c("SF", "SF Mean")
)
}
\arguments{
\item{this_sf}{array of seasonal factors stored as a time series}

\item{y_limit}{Numeric vector of length 2; Range of values you wish the plot to be plotted over.
Default is range of the seasonal factors.}

\item{this_trans}{Logical scalar; indicates if the adjustment was done with a log transform.
Default is TRUE.}

\item{main_title}{Character string; main title of plot.  Default is  \code{'Seasonal Sub-Plots'}.}

\item{sub_title}{Character string; subtitle of plot.  Subtitle not produced if not specified.}

\item{this_xlab}{Character string; label for x-axis of plot.  Default is a blank x-axis.}

\item{do_grid}{Logical scalar; indicates if plots will have grid lines.
Default is no grid lines.}

\item{do_background}{Logical scalar; indicates grey background included in plot.
Default is no grey background;}

\item{this_color_sf}{Character string; color used for seasonal factors.
Default is \code{"darkblue"}.}

\item{this_color_mean}{Character string; color used for means of the seasonal factors.
Default is \code{"darkgrey"}.}

\item{first_year}{Integer scalar; First year used in plot. Default is start of the series.}

\item{add_mean_line}{Logical scalar; indicates if seasonal factor plots will include lines for seasonal means.
Default includes lines for seasonal means.}

\item{this_legend_title}{Character string; indicates title of legend. Default is \code{'Series'}.}

\item{this_legend_text}{Array of character strings; indicates text for each seasonal factor in plot.
Default is \code{c("SF", "SF Mean")}.}
}
\value{
A \code{ggplot} object which generates a plot of the seasonal factors
(and the SI-ratios) grouped by month/quarter.
}
\description{
Generates a special plot of the seasonal factors grouped by month/quarter.
}
\details{
Version 2.1, 9/2/2024
}
\examples{
air_seas    <- 
   seasonal::seas(AirPassengers, arima.model = '(0 1 1)(0 1 1)', 
                  x11 = '')
air_sf      <- seasonal::series(air_seas, 'd10')
plot_air_sf <- plot_sf_series(air_sf,
   add_mean_line = TRUE, 
   main_title = 'Air Passengers Seasonal Sub-Plots',
   this_color_sf = 'darkgreen', 
   this_color_mean = 'lightgreen',
   this_legend_title = 'X-11 Seasonal',
   this_legend_text = c('sf', 'mean'))
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
