% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sf_two.r
\name{plot_sf_two}
\alias{plot_sf_two}
\title{Seasonal factor plot grouped by month/quarter for two series}
\usage{
plot_sf_two(
  this_sf_one = NULL,
  this_sf_two = NULL,
  y_limit = NULL,
  this_trans = TRUE,
  main_title = "Seasonal Sub-Plots",
  sub_title = NULL,
  this_xlab = NULL,
  do_grid = FALSE,
  do_background = FALSE,
  this_color_sf = NULL,
  this_color_mean = NULL,
  this_palette = "Paired",
  first_year = NULL,
  this_legend_title = "SF Plot",
  this_legend_text = c("SF One", "SF Mean One", "SF Two", "SF Mean Two"),
  legend_title_size = 12,
  legend_text_size = 10
)
}
\arguments{
\item{this_sf_one}{array of seasonal factors stored as a time series}

\item{this_sf_two}{array of seasonal factors stored as a time series}

\item{y_limit}{Numeric vector of length 2; Range of values you wish the plot to be plotted over.
Default is range of the seasonal factors.}

\item{this_trans}{Logical scalar; indicates if the adjustment was done with a log transform.
Default is TRUE.}

\item{main_title}{Character string; main title of plot.  Default is  \code{'Seasonal Sub-Plots'}.}

\item{sub_title}{Character string; subtitle of plot.  Subtitle not produced if not specified.}

\item{this_xlab}{Character string; label for x-axis of plot.  Default is a blank x-axis.}

\item{do_grid}{Logical scalar; indicates if plots will have grid lines.
Default is no grid lines.}

\item{do_background}{Logical scalar; indicates grey background included in plot.
Default is no grey background;}

\item{this_color_sf}{Vector of character strings; colors used for seasonal factors. Should be of length two.
Default is \code{NULL}, which indicates that the palette specified
in \code{this_palette} is used to generate colors for this plot.}

\item{this_color_mean}{Vector of character strings; color used for means of the seasonal factors. Should be of length two.
Default is \code{NULL}, which indicates that the palette specified
in \code{this_palette} is used to generate colors for this plot.}

\item{this_palette}{Color used for lines in plot. Default is \code{"Paired"}}

\item{first_year}{Integer scalar; First year used in plot. Default is start of the series.}

\item{this_legend_title}{Character string; indicates title of legend. Default is \code{'Series'}.}

\item{this_legend_text}{Array of character strings; indicates text for each seasonal factor in plot.
Default is \code{c("SF One", "SF Mean One", "SF Two", "SF Mean Two"))}.}

\item{legend_title_size}{integer scalar; Size of the legend title.
Default is \code{12}.}

\item{legend_text_size}{integer scalar; Size of the legend title.
Default is \code{10}.}
}
\value{
A \code{ggplot} object which generates a plot of the seasonal factors
(and the SI-ratios) grouped by month/quarter.
}
\description{
Generates a plot of the seasonal factors grouped by month/quarter
for two adjustments.
}
\details{
Version 1.2, 12/11/2024
}
\examples{
air_seas_x11    <- 
   seasonal::seas(AirPassengers, arima.model = '(0 1 1)(0 1 1)', 
                  x11 = '')
air_sf_x11      <- seasonal::series(air_seas_x11, 'd10')
air_seas_seats  <- 
   seasonal::seas(AirPassengers, arima.model = '(0 1 1)(0 1 1)', 
                  seats.save = "s10")
air_sf_seats    <- seasonal::series(air_seas_seats, 's10')
plot_air_sf_two <- plot_sf_two(air_sf_x11, air_sf_seats,
   main_title = 'Air Passengers Seasonal Sub-Plots',
   this_legend_text = c('X-11 sf', 'X-11 mean', 'SEATS sf', 'SEATS mean'),
   legend_title_size = 10,
   legend_text_size = 8)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
