% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sliding_spans.r
\name{plot_sliding_spans}
\alias{plot_sliding_spans}
\title{Compare spans from the sliding spans analysis}
\usage{
plot_sliding_spans(
  seas_obj = NULL,
  this_series = "sfs",
  main_title = "Plot of Sliding Spans",
  sub_title = NULL,
  do_grid = FALSE,
  do_background = FALSE,
  line_color = NULL,
  this_palette = "Dark2",
  this_guide_legend = "Sliding Spans"
)
}
\arguments{
\item{seas_obj}{\code{seas} object generated from a call of \code{seas} on a single time series
This is a required entry.}

\item{this_series}{Character string; three character code for the type of series to be generated.
Allowed entries are \code{"sfs"} (seasonal factors, default),
\code{"chs"} (period-to-period changes),
\code{"sis"} (indirect seasonal factors),
\code{"cis"} (indirect period-to-period changes).}

\item{main_title}{Title for the plot.
Default is character string \code{'Plot of Sliding Spans'}.}

\item{sub_title}{Subtitle for the plot. Optional entry.}

\item{do_grid}{Logical scalar; indicates if certain plots will have grid lines.
Default is no grid lines.}

\item{do_background}{Logical scalar; indicates grey background included in plot.
Default is no grey background;}

\item{line_color}{Character vector of length 2 to 4 based on the number of spans.
Default is generated from the \code{RColorBrewer} palette \code{"Dark2"}.}

\item{this_palette}{Character string; default \code{RColorBrewer} palette.
Deault is \code{"Dark2"}.}

\item{this_guide_legend}{Title for legend.  Default is \code{"Sliding Spans"}}
}
\value{
A \code{ggplot} object that generates a plot compares the series from
each of the sliding spans generated from a sliding spans analysis
}
\description{
Generates a \code{ggplot} object with a time series plot that compares the series from
each of the sliding spans generated from a sliding spans analysis
}
\details{
Version 1.7, 11/7/2024
}
\examples{
shoes_sspan_seas <- 
   seasonal::seas(blsplotGG::shoes2008, 
                  arima.model = "(0 1 1)(0 1 1)", 
                  transform.function = "log", 
                  forecast.maxlead = 36,
                  slidingspans.save = c('sfs', 'chs'))
p_shoes_spans_sf <- 
    plot_sliding_spans(shoes_sspan_seas, "sfs",
                       main_title = "Spans of Seasonal Factors",
                       sub_title = "US Shoe Sales",
                       this_palette = "Set1")
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
