% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sliding_spans_two.r
\name{plot_sliding_spans_two}
\alias{plot_sliding_spans_two}
\title{Maximum percent difference plot for two sliding spans runs}
\usage{
plot_sliding_spans_two(
  seas_obj_one = NULL,
  seas_obj_two = NULL,
  this_series = "sfs",
  main_title = "Maximum Percent Difference Comparison Plot",
  sub_title = NULL,
  this_x_label = "Time",
  this_y_label = " ",
  do_grid = FALSE,
  do_background = FALSE,
  line_color = "grey",
  cut_color = "red",
  this_legend_title = "Sliding Spans",
  this_legend_entry = c("ss1", "ss2"),
  legend_title_size = 12,
  legend_text_size = 10
)
}
\arguments{
\item{seas_obj_one}{\code{seas} object generated from a call of \code{seas} on a single time series
This is a required entry.}

\item{seas_obj_two}{\code{seas} object generated from a call of \code{seas} on a single time series
This is a required entry.}

\item{this_series}{Character string; three character code for the type of series to be generated.
Allowed entries are \code{"sfs"} (seasonal factors, default),
\code{"chs"} (period-to-period changes),
\code{"sis"} (indirect seasonal factors),
\code{"cis"} (indirect period-to-period changes).}

\item{main_title}{Title for the plot.
Default is character string \code{'Maximum Percent Difference Comparison Plot'}.}

\item{sub_title}{Subtitle for the plot. Optional entry.}

\item{this_x_label}{Label for X-axis.  Default is \code{"Time"}}

\item{this_y_label}{Label for Y-axis.  Default is \code{" "}}

\item{do_grid}{Logical scalar; indicates if certain plots will have grid lines.
Default is no grid lines.}

\item{do_background}{Logical scalar; indicates grey background included in plot.
Default is no grey background;}

\item{line_color}{Color used for lines in the maximum percentage difference plot.
Default is \code{"steelblue"}.}

\item{cut_color}{Color use to show the slidings spans cut off for this type of series.
Default is \code{"red"}.}

\item{this_legend_title}{Character string; indicates title of legend.
Default is \code{'Sliding Spans'}.}

\item{this_legend_entry}{Character array; entries for the legend.
Default is \code{c("ss1", "ss2")}.}

\item{legend_title_size}{integer scalar; Size of the legend title.
Default is \code{12}.}

\item{legend_text_size}{integer scalar; Size of the legend title.
Default is \code{10}.}
}
\value{
A \code{ggplot} object that generates a plot of the maximum percent difference
from a sliding spans analysis of seasonal factors or changes.
}
\description{
Generates a \code{ggplot} object comparing the time series of the maximum percent difference from
the sliding spans analysis of seasonal factors or changes for two sliding spans runs.
The sliding spans analysis for both series should have the same length of spans.
}
\details{
Version 1.5, 12/11/2024
}
\examples{
shoes_sspan_seas <- 
   seasonal::seas(blsplotGG::shoes2008, 
                  arima.model = "(0 1 1)(0 1 1)", 
                  transform.function = "log", 
                  forecast.maxlead = 36,
                  slidingspans.save = c('sfs', 'chs'))

shoes_sspan_x11_seas <- 
   seasonal::seas(blsplotGG::shoes2008, 
                  arima.model = "(0 1 1)(0 1 1)", 
                  transform.function = "log", 
                  forecast.maxlead = 36, 
                  x11.seasonalma = "s3x3",
                  slidingspans.length = 76,
                  slidingspans.numspans = 4,
                  slidingspans.print = 'sfs',
                  slidingspans.save = c('sfs', 'chs'))
p_shoes_maxpct_sf <- 
    plot_sliding_spans_two(shoes_sspan_seas, shoes_sspan_x11_seas, "sfs",
    		main_title = "Maximum Difference of the Seasonal Factors",
         sub_title = "US Retail Shoe Sales",
         line_color = "black", 
         cut_color = "purple",
         this_legend_entry = c("Seats", "X-11"))
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
