% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spectrum.r
\name{plot_spectrum}
\alias{plot_spectrum}
\title{Generate spectrum plot of either the original, seasonally adjusted, irregular, or model residuals.}
\usage{
plot_spectrum(
  seas_obj = NULL,
  this_spectrum = "sp0",
  xaxis_bls = TRUE,
  main_title = "AR Spectrum",
  sub_title = NULL,
  series_name = NULL,
  do_grid = FALSE,
  do_background = FALSE,
  this_color = "darkblue",
  this_median_color = "blue",
  this_freq_color = c("steelblue", "forestgreen"),
  this_peak_color = c("violet", "brown")
)
}
\arguments{
\item{seas_obj}{\code{seas} object generated from a call of \code{seas} on a single time series
This is a required entry.}

\item{this_spectrum}{Character string; three character code for the X-13 spectrum to be generated.
Allowed entries are \code{"sp0"} (modified original series),
\code{"sp1"} (modified X-11 seasonally adjusted series),
\code{"sp2"} (modified X-11 irregular),
\code{"s1s"} (modified SEATS seasonally adjusted series),
\code{"s2s"} (modified SEATS irregular),
\code{"is0"} (modified composite series),
\code{"is1"} (modified indirect seasonally adjusted series),
\code{"is2"} (modified indirect irregular), \code{spr} (model residuals), or
\code{"ser"} (extended residuals).  Default: \code{"sp0"}.}

\item{xaxis_bls}{Logical scalar; indicates if x-axis of spectral plot will be frequency by month
rather than the actual frequencies. Default sets x-axis to frequency by month.}

\item{main_title}{Character string; main title of plot.  Default is  \code{'AR Spectrum'}.}

\item{sub_title}{Character scalar; Description of time series used in \code{seas_obj}.
Used as the subtitle of the plot if specified.}

\item{series_name}{Character scalar; name of the time series used in \code{seas_obj}.
Used as the label of the Y-axis if specified.}

\item{do_grid}{Logical scalar; indicates if certain plots will have grid lines.
Default is no grid lines.}

\item{do_background}{Logical scalar; indicates grey background included in plot.
Default is no grey background;}

\item{this_color}{Character string. Colors used for spectrum in plot.
Default is \code{"darkblue"}.}

\item{this_median_color}{Character string. Colors used for medians of the spectrum.
Default is \code{"blue"}.}

\item{this_freq_color}{Character vector of length 2.
Colors used for seasonal and trading day frequencies, respectively.
Defaults are \code{c("steelblue", "forestgreen")}.}

\item{this_peak_color}{Character vector of length 2.
Colors used for peaks at seasonal and trading day frequencies, respectively.
Defaults are \code{c("violet", "brown")}.}
}
\value{
A \code{ggplot} object which generates a spectrum plot generated by X-13ARIMA-SEATS.
}
\description{
Generate plot of spectrum from X-13ARIMA-SEATS specified by the user.
}
\details{
Version 2.8, 11/6/2024
}
\examples{
air_seas <- seasonal::seas(AirPassengers, arima.model = "(0 1 1)(0 1 1)", x11="", 
                           spectrum.save = "sp2")
plot_air_spectrum <- 
       plot_spectrum(air_seas, this_spectrum = "sp2", 
                     series_name = "AirPassengers",
                     sub_title = "Airline Passengers",
                     this_color = "steelblue", this_median_color = "blue", 
                     this_freq_color = c("darkblue", "darkgreen"), 
                     this_peak_color = c("red", "orange"))
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
