% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_squared_gain.r
\name{plot_squared_gain}
\alias{plot_squared_gain}
\title{Plot of the squared gains for filters generated by an X-13ARIMA-SEATS
SEATS seasonal adjustment run.}
\usage{
plot_squared_gain(
  seas_obj = NULL,
  this_series = "sa",
  main_title = NULL,
  sub_title = NULL,
  this_y_label = "Squared Gain",
  this_x_label = "Cycles per Year",
  do_grid = FALSE,
  do_background = FALSE,
  line_color = NULL,
  this_palette = "Paired",
  this_guide_legend = "Filter"
)
}
\arguments{
\item{seas_obj}{\code{seas} object generated from a call of \code{seas} on a single time series
This is a required entry.}

\item{this_series}{Character string; series for which SEATS produces a squared gain plot,
limited to the seasonally adjusted series (\code{sa}, the default),
or the trend component (\code{trend}).
For other entries, the function will print an error message and return a NULL.}

\item{main_title}{Character string; main title of plot.
A title will be generated if no title is specified.}

\item{sub_title}{Character string; subtitle of plot. There is no default subtitle.}

\item{this_y_label}{Character string; y-axis label for plot.
Default is \code{"Squared Gain"}.}

\item{this_x_label}{Label for X axis. Default is \code{"Cycles per Year"}.}

\item{do_grid}{Logical scalar; indicates if certain plots will have grid lines.
Default is no grid lines.}

\item{do_background}{Logical scalar; indicates grey background included in plot.
Default is no grey background;}

\item{line_color}{Character vector of length two; colors used for the squared gain #'        in the plot.
Default is \code{NULL}, which indicates that the palette specified
in \code{this_palette} is used to generate colors for this plot.}

\item{this_palette}{Color used for lines in plot. Default is \code{"Paired"}}

\item{this_guide_legend}{Title for legend.  Default is \code{"Filter"}}
}
\value{
A \code{ggplot} object which generates a plot of the squared gains for
filters generated by the SEATS seasonal factors.
If SEATS seaonal adjustment (with \code{finite = yes}) not producted,
print out error message and return NULL.
}
\description{
Generate squared gains plot of the concurrent and symmetric SEATS seasonal adjustment
and trend filters from a SEATS adjustment from a \code{seas} object generated by the \code{seasonal}
package.
}
\details{
Version 1.7, 9/5/2024
}
\examples{
shoes_seats_seas <- 
   seasonal::seas(shoes2008, arima.model = "(0 1 1)(0 1 1)", 
                  transform.function = "log", 
                  forecast.maxlead = 36,
                  check.print = c( 'pacf', 'pacfplot' ), 
                  seats.finite = "yes",
                  seats.save = c( 'gac', 'gaf', 'gtc', 'gtf' ) )
p_sa_squared_gain <- plot_squared_gain(shoes_seats_seas, 
	sub_title = "US Shoe Sales",
	this_palette = "Set2")
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{monsell.brian@gmail.com}
}
