% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_table.r
\name{plot_table}
\alias{plot_table}
\title{Plot table from X-13ARIMA-SEATS seasonal adjustment run.}
\usage{
plot_table(
  seas_obj = NULL,
  this_table = NULL,
  main_title = NULL,
  sub_title = NULL,
  this_y_label = NULL,
  y_limit = NULL,
  this_x_label = "Time",
  start_plot = NULL,
  do_grid = FALSE,
  do_background = FALSE,
  draw_recess = FALSE,
  recess_color = "lightgrey",
  recess_sub = TRUE,
  add_outlier = FALSE,
  use_ratio = FALSE,
  ratio_mean = 1,
  this_line_type = "solid",
  line_color = "grey",
  outlier_color = c("red", "blue", "orangered", "green", "steelblue", "blue"),
  outlier_line_type = c("dashed", "dotdash", "dashed", "twodash", "dotdash", "dotdash")
)
}
\arguments{
\item{seas_obj}{\code{seas} object generated from a call of \code{seas} on a single time series
This is a required entry.}

\item{this_table}{Character string; X-13ARIMA-SEATS table name or abbreviation.
If not a valid table name, the function will print an error message
and return a NULL.}

\item{main_title}{Character string; main title of plot.
A title will be generated if no title is specified.}

\item{sub_title}{Character string; subtitle of plot. There is no default subtitle.}

\item{this_y_label}{Character string; y-axis label for plot, if specified.}

\item{y_limit}{Numeric vector of length 2; Range of values for the y-axis.
Default is range of the series specified.}

\item{this_x_label}{Label for X axis. Default is \code{"Time"}.}

\item{start_plot}{Integer vector of length 2; Starting date for plot.
Default is starting date for the time series.}

\item{do_grid}{Logical scalar; indicates if certain plots will have grid lines.
Default is no grid lines.}

\item{do_background}{Logical scalar; indicates grey background included in plot.
Default is no grey background;}

\item{draw_recess}{Logical scalar; indicates if certain plots will have shaded
areas for NBER recession dates.
Default is no recession shading.}

\item{recess_color}{Character string; color used for shading of recession region.
Default is \code{'lightgrey'}.}

\item{recess_sub}{Logical scalar; indicates if x-axis label for recession is
produced for this plot.
Default is x-axis label.}

\item{add_outlier}{Logical scalar; indicates if lines for identified outliers are
included in series plots.
Default is not including lines for identified outliers.}

\item{use_ratio}{Logical scalar; indicates if plots of seasonal factors,
irregular, and residuals are done as ratio plots.
Default has these plots as time series line plots.}

\item{ratio_mean}{Assumed mean value for the ratio.  Default is \code{1.0}}

\item{this_line_type}{Character string; indicates line type of each plot produced.
Default is "solid".}

\item{line_color}{Character string; color used for series in the plot.
Default is \code{'grey'}.}

\item{outlier_color}{Character array of length 6; color used for different
outliers, with the order being \code{'ao', 'ls', 'tc', 'so', 'rp', 'tls'}.
Default is \code{c("red", "blue", 'orangered', "green", "steelblue", "blue")}.}

\item{outlier_line_type}{Character array of length 6; Line type used for different
outliers, with the order being \code{'ao', 'ls', 'tc', 'so', 'rp', 'tls'}.
Default is \code{c('dashed', 'dotdash', 'dashed', 'twodash', 'dotdash', 'dotdash')}.}
}
\value{
A \code{ggplot} object that generates a plot of user-specified series from an
X-13ARIMA-SEATS table.
If series not specified, print out error message and return NULL.
}
\description{
Generate plot of user-specified series from a \code{seas} object generated by the \code{seasonal} package.
}
\details{
Version 2.7, 9/2/2024
}
\examples{
air_seas <- 
  seasonal::seas(AirPassengers, arima.model = "(0 1 1)(0 1 1)", x11="",
                 series.save = 'b1', transform.function = "log",
                 x11.save = "e3")
air_d11_p <- blsplotGG::plot_table(air_seas, "d11", 
         this_y_label = "AirPassengers", 
         main_title = "X-11 Seasonal Adjustment of Airline Passengers",
         sub_title = "Box-Jenkins Airline series",
         do_grid = TRUE, draw_recess = TRUE, 
         use_ratio = FALSE, add_outlier = TRUE, line_color = "darkblue")
air_d16_p <- blsplotGG::plot_table(air_seas, "d16", 
         this_y_label = "AirPassengers", 
         main_title = "X-11 Seasonal Adjustment of Airline Passengers",
         do_grid = FALSE, draw_recess = TRUE, 
         use_ratio = TRUE, add_outlier = TRUE, line_color = "steelblue")
air_e3_p <- blsplotGG::plot_table(air_seas, "e3",
         this_y_label = "AirPassengers",
         main_title = "X-11 Seasonal Adjustment (Extreme Adjusted) of Airline Passengers",
         do_grid = FALSE, draw_recess = TRUE,
         use_ratio = FALSE, add_outlier = TRUE, line_color = "steelblue")
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
