% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_two_sa.r
\name{plot_two_sa}
\alias{plot_two_sa}
\title{Compare two seasonal adjustments}
\usage{
plot_two_sa(
  this_sa_one = NULL,
  this_sa_two = NULL,
  this_ori = NULL,
  main_title = "Compare X-11 and SEATS",
  sub_title = NULL,
  this_x_label = "Time",
  this_y_label = " ",
  this_sa_text = c("X-11", "SEATS"),
  do_grid = FALSE,
  do_background = FALSE,
  line_color = NULL,
  this_palette = "Dark2",
  this_guide_legend = "Series"
)
}
\arguments{
\item{this_sa_one}{Time series of the X-11 seasonal adjustment.
This is a required entry.}

\item{this_sa_two}{Time series of the SEATS seasonal adjustment.
This is a required entry.}

\item{this_ori}{Time series of the original series. Optional entry.}

\item{main_title}{Title for the plot.
Default is character string \code{'Comparison of Seasonal Adjustments'}.}

\item{sub_title}{Subtitle for the plot. Optional entry.}

\item{this_x_label}{Label for X-axis.  Default is \code{"Time"}}

\item{this_y_label}{Label for Y-axis.  Default is \code{" "}}

\item{this_sa_text}{Labels for different seasonal adjustments.
Default is \code{c('X-11', 'SEATS')}}

\item{do_grid}{Logical scalar; indicates if certain plots will have grid lines.
Default is no grid lines.}

\item{do_background}{Logical scalar; indicates grey background included in plot.
Default is no grey background;}

\item{line_color}{Character scalar; color used for plot.
User should specify one color for each column of the matrix specified.
Default is the \code{RColorBrewer} palette \code{"Dark2"}.}

\item{this_palette}{Character string; default \code{RColorBrewer} palette.
Deault is \code{"Dark2"}.}

\item{this_guide_legend}{Title for legend.  Default is \code{"Series"}}
}
\value{
A \code{ggplot} object that generates a plot comparing two seasonal adjustments, trend, or factors.
}
\description{
Generates a \code{ggplot} object with a time series plot that compares two seasonal adjustments
of the same series, optionally including the original series.
}
\details{
Version 2.3, 8/26/2024
}
\examples{
ukgas_x11_seas   <- 
   seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)", 
                  transform.function = "log", forecast.maxlead = 20,
                  x11.seasonalma = "s3x5", 
                  check.print = c( 'pacf', 'pacfplot' ))
ukgas_seats_seas <- 
   seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)", 
                  transform.function = "log", forecast.maxlead = 20,
                  check.print = c( 'pacf', 'pacfplot' ))
ukgas_x11_sa     <- seasonal::final(ukgas_x11_seas)
ukgas_seats_sa   <- seasonal::final(ukgas_seats_seas)
ukgas_seats_v_x11_p <- 
    plot_two_sa(this_sa_one = ukgas_x11_sa, this_sa_two = ukgas_seats_sa,
                       main_title = "UK Gas Seasonal Adjustments",
                       sub_title = "X-11 - Blue, SEATS - Green",
                       line_color = c("steelblue", "forestgreen"))
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
