% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_two_sa_facet.r
\name{plot_two_sa_facet}
\alias{plot_two_sa_facet}
\title{Compare two seasonal adjustments in a facet plot}
\usage{
plot_two_sa_facet(
  this_sa_one = NULL,
  this_sa_two = NULL,
  this_ori = NULL,
  main_title = "Compare X-11 and SEATS",
  sub_title = NULL,
  this_x_label = "Time",
  this_y_label = " ",
  this_sa_text = c("X-11", "SEATS"),
  line_color = "steelblue"
)
}
\arguments{
\item{this_sa_one}{Time series of the first seasonal adjustment.
This is a required entry.}

\item{this_sa_two}{Time series of the second seasonal adjustment.
This is a required entry.}

\item{this_ori}{Time series of the original series. Optional entry.}

\item{main_title}{Title for the plot. Default is character string \code{'Comparison of Seasonal Adjustments'}.}

\item{sub_title}{Subtitle for the plot. Optional entry.}

\item{this_x_label}{Label for X-axis.  Default is \code{"Time"}}

\item{this_y_label}{Label for Y-axis.  Default is \code{" "}}

\item{this_sa_text}{Labels for different seasonal adjustments.  Default is \code{c('X-11', 'SEATS')}}

\item{line_color}{Color used for lines in plot.  Default is \code{"steelblue"}.}
}
\value{
A \code{ggplot} object that generates a facet plot comparing two seasonal adjustments,
trends, or factors.
}
\description{
Generates a \code{ggplot} object with a time series facet plot that compares two s
easonal adjustments of the same series, optionally including the original series.
}
\details{
Version 2.1, 5/6/2024
}
\examples{
ukgas_x11_seas   <- 
   seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)", 
                  transform.function = "log", forecast.maxlead = 20,
                  x11.seasonalma = "s3x5", 
                  check.print = c( 'pacf', 'pacfplot' ))
ukgas_seats_seas <- 
   seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)", 
                  transform.function = "log", forecast.maxlead = 20,
                  check.print = c( 'pacf', 'pacfplot' ))
ukgas_x11_sa     <- seasonal::final(ukgas_x11_seas)
ukgas_seats_sa   <- seasonal::final(ukgas_seats_seas)
ukgas_two_sa_facet_p <- 
    plot_two_sa_facet(this_sa_one = ukgas_x11_sa, this_sa_two = ukgas_seats_sa,
                      main_title = "UK Gas Seasonal Adjustments",
                      line_color = "forestgreen")
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{monsell.brian@gmail.com}
}
