% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_x11_and_seats.r
\name{plot_x11_and_seats}
\alias{plot_x11_and_seats}
\title{Compare X-11 and SEATS seasonal adjustment}
\usage{
plot_x11_and_seats(
  this_x11 = NULL,
  this_seats = NULL,
  this_ori = NULL,
  main_title = "Compare X-11 and SEATS",
  sub_title = NULL,
  this_x_label = "Time",
  this_y_label = " ",
  do_grid = FALSE,
  do_background = FALSE,
  line_color = NULL,
  this_palette = "Dark2",
  this_guide_legend = "Series"
)
}
\arguments{
\item{this_x11}{Time series of the X-11 seasonal adjustment.
This is a required entry.}

\item{this_seats}{Time series of the SEATS seasonal adjustment.
This is a required entry.}

\item{this_ori}{Time series of the original series. Optional entry.}

\item{main_title}{Title for the plot.
Default is character string \code{'Comparison of X-11 and SEATS Seasonal Adjustments'}.}

\item{sub_title}{Subtitle for the plot. Optional entry.}

\item{this_x_label}{Label for X-axis.  Default is \code{"Time"}}

\item{this_y_label}{Label for Y-axis.  Default is \code{" "}}

\item{do_grid}{Logical scalar; indicates if certain plots will have grid lines.
Default is no grid lines.}

\item{do_background}{Logical scalar; indicates grey background included in plot.
Default is no grey background;}

\item{line_color}{Character vector of length 2 (if \code{this_ori} is not specified)
or 3 (if \code{plot_ori} is specified); color used for lines in the plot,
in the order of seasonally adjusted series, trend, original series.
Default is generated from the \code{RColorBrewer} palette \code{"Dark2"}.}

\item{this_palette}{Character string; default \code{RColorBrewer} palette.
Deault is \code{"Dark2"}.}

\item{this_guide_legend}{Title for legend.  Default is \code{"Series"}}
}
\value{
A \code{ggplot} object that generates a plot comparing an X-11 and SEATS
seasonal adjustment, trend, or factors.
}
\description{
Generates a \code{ggplot} object with a time series plot that compares an X-11 and SEATS
seasonal adjustment, optionally including the original series.
}
\details{
Version 4.3, 11/6/2024
}
\examples{
ukgas_x11_seas   <- 
   seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)", 
                  transform.function = "log", forecast.maxlead = 20,
                  x11.seasonalma = "s3x5", 
                  check.print = c( 'pacf', 'pacfplot' ))
ukgas_seats_seas <- 
   seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)", 
                  transform.function = "log", forecast.maxlead = 20,
                  check.print = c( 'pacf', 'pacfplot' ))
ukgas_x11_sa     <- seasonal::final(ukgas_x11_seas)
ukgas_seats_sa   <- seasonal::final(ukgas_seats_seas)
ukgas_seats_v_x11_p <- 
    plot_x11_and_seats(this_x11 = ukgas_x11_sa, this_seats = ukgas_seats_sa,
                       main_title = "UK Gas Seasonal Adjustments",
                       sub_title = "X-11 - Blue, SEATS - Green",
                       line_color = c("steelblue", "forestgreen"))
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
