% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_x11_and_seats_facet.r
\name{plot_x11_and_seats_facet}
\alias{plot_x11_and_seats_facet}
\title{Compare X-11 and SEATS seasonal adjustments in a facet plot}
\usage{
plot_x11_and_seats_facet(
  this_x11 = NULL,
  this_seats = NULL,
  this_ori = NULL,
  main_title = "Compare X-11 and SEATS",
  sub_title = NULL,
  this_x_label = "Time",
  this_y_label = " ",
  line_color = "steelblue",
  remove_legend = FALSE
)
}
\arguments{
\item{this_x11}{Time series of the X-11 seasonal adjustment.
This is a required entry.}

\item{this_seats}{Time series of the SEATS seasonal adjustment.
This is a required entry.}

\item{this_ori}{Time series of the original series. Optional entry.}

\item{main_title}{Title for the plot. Default is character string \code{'Comparison of X-11 and SEATS Seasonal Adjustments'}.}

\item{sub_title}{Subtitle for the plot. Optional entry.}

\item{this_x_label}{Label for X-axis.  Default is \code{"Time"}}

\item{this_y_label}{Label for Y-axis.  Default is \code{" "}}

\item{line_color}{Color used for lines in plot.  Default is \code{"steelblue"}.}

\item{remove_legend}{Logical scalar; if TRUE, plot legend will be removed.
Default is FALSE.}
}
\value{
A \code{ggplot} object that generates a facet plot comparing an X-11 and SEATS
seasonal adjustment, trend, or factor.
}
\description{
Generates a \code{ggplot} object with a time series facet plot that compares an X-11 and SEATS
seasonal adjustment, optionally including the original series.
}
\details{
Version 6.3, 9/9/2024
}
\examples{
ukgas_x11_seas   <- 
   seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)", 
                  transform.function = "log", forecast.maxlead = 20,
                  x11.seasonalma = "s3x5", 
                  check.print = c( 'pacf', 'pacfplot' ))
ukgas_seats_seas <- 
   seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)", 
                  transform.function = "log", forecast.maxlead = 20,
                  check.print = c( 'pacf', 'pacfplot' ))
ukgas_x11_sa     <- seasonal::final(ukgas_x11_seas)
ukgas_seats_sa   <- seasonal::final(ukgas_seats_seas)
ukgas_seats_v_x11_facet_p <- 
    plot_x11_and_seats_facet(this_x11 = ukgas_x11_sa, this_seats = ukgas_seats_sa,
                       main_title = "UK Gas Seasonal Adjustments",
                       line_color = "forestgreen")
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{monsell.brian@gmail.com}
}
