% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_outlier.R
\name{proc_outlier}
\alias{proc_outlier}
\title{Extract dates from outlier text}
\usage{
proc_outlier(this_outlier = NULL, this_freq = 12, add_type = TRUE)
}
\arguments{
\item{this_outlier}{Character string; outlier regressor.
This is a required entry.}

\item{this_freq}{integer scalar; time series frequency. Default is \code{12}.}

\item{add_type}{logical scalar; determines if type of outlier is added to the output.
Default is \code{TRUE}.}
}
\value{
List of either year and month/quarter of outlier, or year and month/quarter of
start and end of outlier
}
\description{
Process name of outlier regressor to extract the dates associated with the outlier
}
\details{
Version 2.1, 5/2/2024
}
\examples{
air_seas <- 
   seasonal::seas(AirPassengers, x11="", slidingspans = "", 
                     transform.function = "log", arima.model = "(0 1 1)(0 1 1)", 
                     regression.aictest = 'td', forecast.maxlead=36, 
                     check.print = c( "pacf", "pacfplot" ))
this_auto_outlier <- get_auto_outlier_string(air_seas)
this_outlier      <- proc_outlier(this_auto_outlier)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{monsell.brian@gmail.com}
}
