% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_subplot.r
\name{seasonal_subplot}
\alias{seasonal_subplot}
\title{Seasonal sub-plot}
\usage{
seasonal_subplot(
  this_sf = NULL,
  this_sf_range = NULL,
  main_title = "Seasonal Subplot",
  sub_title = NULL,
  this_x_label = "Month",
  this_y_label = "",
  subplot_color = "steelblue"
)
}
\arguments{
\item{this_sf}{Time series of seasonal factors from X-11 or SEATS}

\item{this_sf_range}{Range of values you wish the plot to be plotted over. Default is range of the series.}

\item{main_title}{Title for the plot. Default is character string \code{'Ratio Plot'}.}

\item{sub_title}{Subtitle for the plot. Default is \code{NULL}.}

\item{this_x_label}{Label for X axis. Default is \code{"Time"}.}

\item{this_y_label}{Label for Y axis. Default is \code{"Ratio"}.}

\item{subplot_color}{Color used for lines in ratio plot.  Default is \code{"steelblue"}.}
}
\value{
A \code{ggplot} object that generates a ratio plot.
}
\description{
Generates a seasonal sub-plot from a ts object of seasonal (or combined adjustment) factors
}
\details{
Version 2.0, 5/6/2024
}
\examples{
ukgas_x11_seas <- 
   seasonal::seas(UKgas, series.period = 4, arima.model = "(0 1 1)(0 1 1)", 
                  transform.function = "log", forecast.maxlead = 20,
                  x11.seasonalma = "s3x5", 
                  check.print = c( 'pacf', 'pacfplot' ))
ukgas_x11_sf     <- seasonal::series(ukgas_x11_seas, "d10")
p_ukgas_sf_sub   <- 
    seasonal_subplot(ukgas_x11_sf, 
                     main_title = "UK Gas X-11 Seasonal Factors",
                     sub_title = "Seasonal Filter: 3x5", 
                     this_x_label = "Quarter")
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{monsell.brian@gmail.com}
}
