% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoes2008.r
\docType{data}
\name{shoes2008}
\alias{shoes2008}
\title{Retail sales of shoes, 2008}
\format{
Retail sales of shoes ending in April of 2008
}
\usage{
shoes2008
}
\description{
A time series object containing retail sales of shoes
}
\keyword{datasets}
