% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_vector.r
\name{update_vector}
\alias{update_vector}
\title{Update vector.}
\usage{
update_vector(this_series = NULL, this_num = NULL)
}
\arguments{
\item{this_series}{Original time series. This is a required entry.}

\item{this_num}{Length of updated series. Must be more than the length of \code{this_series}.
This is a required entry.}
}
\value{
An updated vector of length \code{this_num} augmented with the first value of the
input series.
}
\description{
Fill unspecified elements of a vector with the first element of the input series
}
\details{
Version 2.3, 5/25/2023
}
\examples{
this_vector <- c(1,2)
updated_vector <- update_vector(this_vector, 4)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
