% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/what_spectrum.r
\name{what_spectrum}
\alias{what_spectrum}
\title{What spectrum is plotted}
\usage{
what_spectrum(this_spectrum = "sp0", use_title_case = FALSE)
}
\arguments{
\item{this_spectrum}{Character string; three character code for the X-13 spectrum to be generated.
Allowed entries are \code{"sp0"} (modified original series),
\code{"sp1"} (modified X-11 seasonally adjusted series),
\code{"sp2"} (modified X-11 irregular),
\code{"s1s"} (modified SEATS seasonally adjusted series),
\code{"s2s"} (modified SEATS irregular),
\code{"is0"} (modified composite series),
\code{"is1"} (modified indirect seasonally adjusted series),
\code{"is2"} (modified indirect irregular), \code{spr} (model residuals), or
\code{"ser"} (extended residuals).  Default: \code{"sp0"}.}

\item{use_title_case}{Logical scalar; convert string to title case. Default is FALSE.}
}
\value{
Text for spectrum associated with code used in \code{plot_spectrum} function.
If improper value set for \code{this_spectrum}, function will return \code{NULL}.
}
\description{
Return which spectrum plot is generated based on the three character code used by the \code{plot_spectrum} function.
}
\details{
Version 1.5, 5/6/2024
}
\examples{
sub_title <- what_spectrum("sp2", use_title_case = TRUE)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
