% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xt_data_list.r
\docType{data}
\name{xt_data_list}
\alias{xt_data_list}
\title{US Building Permits, One Family Units}
\format{
A list object with 4 time series elements:
\describe{
\item{mw1u}{Midwest one family building permits}
\item{ne1u}{Northeast one family building permits}
\item{so1u}{South one family building permits}
\item{we1u}{West one family building permits}
}
}
\usage{
xt_data_list
}
\description{
A list object of US One family Building Permits for four regions expressed as time series objects that end in October, 2006
}
\keyword{datasets}
