% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_year_over_year.r
\name{plot_year_over_year}
\alias{plot_year_over_year}
\title{Plot year over year plot}
\usage{
plot_year_over_year(
  this_series = NULL,
  main_title = NULL,
  sub_title = NULL,
  this_y_label = NULL,
  y_limit = NULL,
  this_x_label = NULL,
  start_plot = NULL,
  do_grid = FALSE,
  do_background = FALSE,
  line_color = NULL,
  this_palette = "Paired",
  detrend_series = FALSE,
  detrend_lowess = FALSE
)
}
\arguments{
\item{this_series}{Numeric matrix; columns of time series object to be plotted.}

\item{main_title}{Character string; main title of plot.
The default title is the name of the series passed to this function.}

\item{sub_title}{Character string; subtitle of plot. There is no default subtitle.}

\item{this_y_label}{Character string; y-axis label for plot, if specified.}

\item{y_limit}{Numeric vector of length 2; Range of values you wish the plot to be plotted over.
Default is range of the series specified.}

\item{this_x_label}{Label for X axis. Default is \code{"Month"} or \code{"Quarter"}.}

\item{start_plot}{Integer vector of length 2; Starting date for plot.
Default is starting date for the time series.}

\item{do_grid}{Logical scalar; indicates if plots will have grid lines.
Default is no grid lines.}

\item{do_background}{Logical scalar; indicates grey background included in plot.
Default is no grey background.}

\item{line_color}{Character scalar; color used for plot.
User should specify one color for each column of the matrix specified.
Default is the \code{RColorBrewer} palette \code{"Paired"}.}

\item{this_palette}{Character string; default \code{RColorBrewer} palette.
Default is \code{"Paired"}.}

\item{detrend_series}{Logical scalar; indicates if the series plotted is to be detrended.
Default is the original series is plotted.}

\item{detrend_lowess}{Logical scalar; indicates lowess is used to generate the trend
used to detrend the series. Default is loess is not used.}
}
\value{
Generate \code{ggplot} object generating a year to year plot of a time series object.
If time series object not specified, print out error message and return NULL.
}
\description{
Generate year over year plot of a user-specified ts object.
}
\details{
Version 3.1, 9/25/2024
}
\examples{
this_yyplot <- 
    plot_year_over_year(AirPassengers, this_y_label = "Air", this_palette = "Dark2")
this_yyplot_detrend <- 
    plot_year_over_year(AirPassengers, this_y_label = "Air", this_palette = "Dark2",
                        detrend_series = TRUE, detrend_lowess = TRUE)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
